% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbitr_helpers.R
\name{add_spaces}
\alias{add_spaces}
\title{Add spaces}
\usage{
add_spaces(my_text, to)
}
\arguments{
\item{my_text}{A character vector of text to be justified.}

\item{to}{A single-element character vector with allowed values of either
'head' or 'tail' to indicate on which side spaces should be added.}
}
\value{
A character vector of text with spaces added to give each entry an
equal number of characters.
}
\description{
Justify table columns by adding spaces.
}
\details{
The function \code{add_spaces()} is used to make justified tables with a
monospaced font. The parameter \code{my_text} should contain a vector of text
for a single column, with each row as a separate element. For right-
justified text, set the \code{to} parameter to 'head'. This will add enough
space to the front of each row to make all entries have an equal number of
characters. Similarly, for left-justified text, set the \code{to} parameter to
'tail'.
}
