% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbitr_helpers.R
\name{add_x_intercepts}
\alias{add_x_intercepts}
\title{Add x-intercepts}
\usage{
add_x_intercepts(dataframe, x_name, y_name)
}
\arguments{
\item{dataframe}{A 2-column data frame with numeric columns, where each row
is a point in the x-y plane.}

\item{x_name}{A single-element character vector with the name of the column
containing the x-coordinates.}

\item{y_name}{A single-element character vector with the name of the column
containing the y-coordinates.}
}
\value{
A new data frame consisting of the original data frame with new rows
added for each intersection with the x-axis.
}
\description{
Given a data frame of x-y coordinates, find the x-intercepts of the line
segments that connect neighboring pairs of points, and add the x-intercepts
to the data frame.
}
\details{
The input data frame is expected to have two columns, with each row
being the Cartesian coordinates of a point in the x-y plane. If the points
are arranged in order of increasing x values and connected by line
segments, \code{add_x_intercepts} will find the points where the segments
intersect the x-axis and add those points to the data frame.
}
