% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_to_fen.R
\name{board_to_fen}
\alias{board_to_fen}
\title{Convert a chessboard state to a FEN string}
\usage{
board_to_fen(board)
}
\arguments{
\item{board}{A list containing an 8x8 matrix of the board, whose trun it is,
castling rights, en passant target square, halfmove clock, and fullmove
number.}
}
\value{
A FEN string representing the board.
}
\description{
The purpose of \code{board_to_fen()} is simply to reverse the operation of
\code{\link[=fen_to_board]{fen_to_board()}}. As such, \code{board_to_fen()} accepts as input a list
of game data (\code{board}) in the format produced by \code{fen_to_board()}.
}
\details{
The\code{board} argument is a named list containing the same data as a FEN string,
but in a format that is more accessible in R:

\verb{$board} An 8x8 character matrix representing the chessboard, with lowercase
letters for the black pieces and uppercase for white.
\verb{$to_move}: Whose turn it is. Either 'w' or 'b'.
\verb{$castling_rights}: Which castling privileges remain. Can be any combination
of the characters 'K', 'Q', 'k', 'q',  in order from uppercase to lowercase
and then king before queen to show queenside and kingside castling rights for
white or black, or '-' if no castling rights remain.
\verb{$ep_target}: If the previous move was a two-square pawn advance,
\verb{$ep_target} will give the name of the square where an enemy pawn capturing
\emph{en passant} would land, even if no enemy pawn can currently capture
\emph{en passant}. Otherwise, the value is '-'.
\verb{$halfmove_clock}: An integer for keeping track of the number of half moves
since the last pawn move or capture. Used to detect 50 move rule draws.
\verb{$fullmove_number}: An integer showing the current move number. Increases by
one after every black move.

See
\href{http://www.saremba.de/chessgml/standards/pgn/pgn-complete.htm#c16.1}{section 16.1 of the PGN specification}
for detailed information about the FEN format. Also see the documentation for
\code{\link[=fen_to_board]{fen_to_board()}} for more details about the board format.
}
\examples{
board <- fen_to_board()
board_to_fen(board)
}
\seealso{
\itemize{
\item \code{\link[=fen_to_board]{fen_to_board()}} To convert a FEN string to a board matrix.
\item \code{\link[=print_board]{print_board()}} Prints the board with rows in the correct order.
}
}
