% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_to_int.R
\name{board_to_int}
\alias{board_to_int}
\title{Convert chessboard characters to integers}
\usage{
board_to_int(board)
}
\arguments{
\item{board}{A character matrix representing a chessboard.}
}
\value{
An integer matrix with the same dimensions as the input, where chess
pieces have been converted to integers.
}
\description{
This function takes an 8x8 character matrix of a chessboard and converts
letter representations of pieces into integers, with black pieces indicated
using a - sign.
}
\details{
1 = pawn
2 = bishop
3 = knight
4 = rook
5 = queen
6 = king
}
\examples{
board <- fen_to_board()
int_board <- board_to_int(board$board)

}
