% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgn_data_manipulators.R
\name{calculate_move_times}
\alias{calculate_move_times}
\title{Convert clock data to move times}
\usage{
calculate_move_times(clock, increment, color)
}
\arguments{
\item{clock}{An integer vector of chess clock times at the end of each half
move (in seconds).}

\item{increment}{A single-element integer vector of the increment time (in
seconds).}

\item{color}{A character vector indicating which side to calculate (either
black or white).}
}
\value{
A numeric vector of the move times (if present) in seconds. The
vector starts at the second move (the first move is zero seconds).
}
\description{
Clock data (in seconds) is converted into move times (in seconds).
}
\details{
Clock times are typically obtained from \code{rbitr::get_clocks}, and
increment times from \code{rbitr::get_increments}. The move times start from the
second move (the first move is taken to be zero seconds).
}
\note{
Data from lichess.org may be subject to errors due to unrecorded
additions of time from take backs, or from the opponent granting more time.
If either of these events occur in a given move, then the calculated move
time will be shorter than the actual move time. This error can only be
detected when it results in a negative move time, in which case the value is
replaced with NA.
}
\examples{
clock <- c(900, 900, 888, 878, 878, 858)
calculate_move_times(clock, 8, 'white')
calculate_move_times(clock, 8, 'black')
}
\seealso{
Other pgn data manipulators: 
\code{\link{get_clocks}()},
\code{\link{get_increments}()}
}
\concept{pgn data manipulators}
