% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/can_bishop_move.R
\name{can_bishop_move}
\alias{can_bishop_move}
\alias{bishop_can_move}
\title{Check if a bishop has any legal moves}
\usage{
can_bishop_move(board, coordinates, color = board$to_move)

bishop_can_move(board, coordinates, color = board$to_move)
}
\arguments{
\item{board}{A list containing the current state of the chess board.}

\item{coordinates}{A vector of length 2 representing the coordinates of the
bishop on the board. The first element is the row number and the second
element is the column number.}

\item{color}{An optional character vector giving the color of the bishop.
Defaults to the color of the player whose turn it is to move.}
}
\value{
A logical value indicating whether the bishop has any legal moves
(TRUE) or not (FALSE).
}
\description{
This function checks if a bishop at a given position on a chess board has any
legal moves.
}
\details{
See \code{\link[=fen_to_board]{fen_to_board()}} for details of the chess board format.
This function is used by other functions when testing for stalemate.
\code{can_bishop_move()} is an alias for \code{bishop_can_move()}.
}
\examples{
board <- fen_to_board() # Starting position
coordinates <- c(1, 3)  # Coordinates of the white bishop in the starting position
can_bishop_move(board, coordinates)
}
