% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/can_knight_move.R
\name{can_knight_move}
\alias{can_knight_move}
\alias{knight_can_move}
\title{Check if a knight has any legal moves}
\usage{
can_knight_move(board, coordinates, color = board$to_move)

knight_can_move(board, coordinates, color = board$to_move)
}
\arguments{
\item{board}{A list containing the current state of the chess board.}

\item{coordinates}{A vector of length 2 representing the coordinates of the
knight on the board. The first element is the row number and the second
element is the column number.}

\item{color}{An optional character vector giving the color of the knight.
Defaults to the color of the player whose turn it is to move.}
}
\value{
A logical value indicating whether the knight has any legal moves
(TRUE) or not (FALSE).
}
\description{
This function checks if a knight at a given position on a chess board has any
legal moves.
}
\details{
See \code{\link[=fen_to_board]{fen_to_board()}} for details of the chess board format.
This function is used by other functions when testing for stalemate.
\code{can_knight_move()} is an alias for \code{knight_can_move()}.
}
\examples{
board <- fen_to_board() # Starting position
coordinates <- c(1, 2)  # Coordinates of the white knight in the starting position
can_knight_move(board, coordinates)
}
