% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_movetext.R
\name{clean_movetext}
\alias{clean_movetext}
\title{Remove comments and annotations from movetext.}
\usage{
clean_movetext(movetext)
}
\arguments{
\item{movetext}{A character vector of pgn movetext fields.}
}
\value{
A character vector of clean movetext.
}
\description{
The \href{http://www.saremba.de/chessgml/standards/pgn/pgn-complete.htm#c8.2}{pgn specification}
allows comments, annotations, game termination markers, and other formatting
oddities that may not play well with all software. The \code{clean_movetext()}
function removes everything but the moves and move numbers.
}
\examples{
clean_movetext('1. e4! {Best by test.} e5')
}
