% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_moves.R
\name{convert_to_lan}
\alias{convert_to_lan}
\title{convert_to_san}
\usage{
convert_to_lan(movetext)
}
\arguments{
\item{movetext}{A character vector of movetext in SAN format.}
}
\value{
A character vector of movetext in LAN format.
}
\description{
Internal helper function for \code{get_moves()}. This is a wrapper around
\code{bigchess::san2lan()} to deal with empty movetext. When movetext == '', it
returns ' ', which in \code{get_moves()} gets passed to \code{strsplit()} that will
convert it to ''.
}
