% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cram_positionlog.R
\name{convert_to_numeric}
\alias{convert_to_numeric}
\title{Convert character columns containing only numeric data to numeric}
\usage{
convert_to_numeric(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A converted data frame
}
\description{
This is a helper function for \code{cram_positionlog()}. It assumes that the data
frame can only have numeric or character columns (i.e., no factors). It
looks for columns where every string contains numeric data, and coverts the
them to numeric. If the column contains any non-numeric or mixed values, it
will not be converted.
}
