% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_ply.R
\name{count_ply}
\alias{count_ply}
\title{Count the number of ply in a chess game}
\usage{
count_ply(movetext)
}
\arguments{
\item{movetext}{A character string containing the movetext of a chess game.}
}
\value{
An integer representing the number of ply in the game.
}
\description{
This function takes the movetext of a chess game as input and returns the
number of ply (half-moves) in the game.
}
\examples{
count_ply("1. e4 e5 2. Nf3 Nc6")
}
