% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_ply.R
\name{count_ply}
\alias{count_ply}
\alias{count_halfmoves}
\title{Count the number of ply or halfmoves in a chess game}
\usage{
count_ply(movetext)

count_halfmoves(movetext)
}
\arguments{
\item{movetext}{A character string containing the movetext of a chess game in
either standard algebraic notation (SAN) or
\href{http://wbec-ridderkerk.nl/html/UCIProtocol.html}{UCI-format} long
algebraic notation (LAN).}
}
\value{
\itemize{
\item \code{count_ply()} returns an integer of the number of ply in the game.
\item \code{count_halfmoves()} returns an integer of the number of half-moves in the game.
}
}
\description{
The functions \code{count_ply()} and \code{count_halfmoves()} take the movetext of a
chess game as input and return either the number of ply (positions) or the
number of half-moves in the game.
}
\details{
Repeated positions or repeated half moves are included in the count.
The number of half-moves is always one less than the number of ply.
}
\examples{
count_ply("1. e4 e5 2. Nf3 Nc6")
count_halfmoves("1. e4 e5 2. Nf3 Nc6")
}
