% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_tools.R
\name{cull_pvs}
\alias{cull_pvs}
\title{Extract principal variations (PVs) from UCI chess engine output.}
\usage{
cull_pvs(ucilog, depth = NULL)
}
\arguments{
\item{ucilog}{A character vector of engine output.}

\item{depth}{(Optional, default = NULL) A single-element integer vector
indicating which search depth to return. The value must not be less than 1
or greater than the maximum depth reported by the engine. A value of NULL
returns scores for the maximum depth}
}
\value{
A character vector of the principal variations in LAN format.
}
\description{
Parse the output of a uci-compatible chess engine for the principal
variations.
}
\details{
The principal variations for the deepest available search depth are
returned by default. See wbec-ridderkerk.nl/html/UCIProtocol.html for
details of UCI engine output.
}
\examples{
library(bigchess)
position <- 'e2e4 g7g5 b1c3'
engine_path <- '../../stockfish_13_win_x64_bmi2.exe'
engine <- uci_engine(engine_path)
ucilog <- evaluate_position(position, engine, n_pv = 2, depth = 2)
cull_pvs(ucilog)
}
