% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_tools.R
\name{cull_scores}
\alias{cull_scores}
\title{Extract scores from UCI chess engine output.}
\usage{
cull_scores(ucilog, depth = NULL)
}
\arguments{
\item{ucilog}{A character vector of engine output.}

\item{depth}{(Optional, default = maximum) A single-element integer vector
indicating which search depth to return. The value must not be less than 1
or greater than the maximum depth reported by the engine.}
}
\value{
A character vector of the score (or mate in x) for each PV.
}
\description{
Parse the output of a uci-compatible chess engine for the engine's
evaluation of the position.
}
\details{
The output is left in character format rather than converting to
numeric to preserve any results that are 'mate in x'. The score for the
deepest available search depth is returned by default. See
wbec-ridderkerk.nl/html/UCIProtocol.html for details of UCI engine output.
}
\examples{
library(bigchess)
position <- 'e2e4 g7g5 b1c3'
engine_path <- '../../stockfish_13_win_x64_bmi2.exe'
engine <- uci_engine(engine_path)
ucilog <- evaluate_position(position, engine, n_pv = 2, depth = 2)
cull_scores(ucilog, depth = 2)
}
