% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_games.R
\name{evaluate_games}
\alias{evaluate_games}
\title{Evaluate Games}
\usage{
evaluate_games(games, engine, n_pv, depth)
}
\arguments{
\item{games}{A list of games, where each game is a list of chess positions.}

\item{engine}{An engine handler generated by the bigchess function
\code{uci_engine()}.}

\item{n_pv}{A single-element integer vector of the desired number of
principal variations.}

\item{depth}{A single-element integer vector of the desired search depth, in
plies.}
}
\value{
A recursive list of ucilogs.
}
\description{
Evaluate Games
}
\examples{
games <- list(c('', 'e2e4', 'e2e4 g7g5', 'e2e4 g7g5 b1c3',
                'e2e4 g7g5 b1c3 f7f5', 'e2e4 g7g5 b1c3 f7f5 d1h5'))
# Modify the path as required for your engine location & operating system
engine <- bigchess::uci_engine('//stockfish_13_win_x64_bmi2.exe')
evaluate_games(games, engine, n_pv = 1, depth = 1)
bigchess::uci_quit(engine)
}
