% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_pgn.R
\name{evaluate_pgn}
\alias{evaluate_pgn}
\title{Evaluate each chess game in a pgn file.}
\usage{
evaluate_pgn(pgn_path, engine, n_pv, depth)
}
\arguments{
\item{pgn_path}{A single-element character vector of the path to the pgn.}

\item{engine}{An engine handler generated by the
\href{https://github.com/rosawojciech/bigchess}{bigchess} function
\code{uci_engine()}.}

\item{n_pv}{A single-element integer vector of the desired number of
principal variations.}

\item{depth}{A single-element integer vector of the desired search depth, in
plies.}
}
\value{
A list. Each element of the list corresponds to a game in the pgn,
and consists of a sublist containing the engine output for each position in
that game.
}
\description{
Evaluate each chess game in a pgn file.
}
\details{
A wrapper for rbitr's \code{evaluate_game()} function that evaluates each
game in the specified pgn file. Note that this can take a very long time
for pgn files with a lot of games, and even longer when searching high
numbers of principal variations and/or deep evaluations.
}
\examples{
library(bigchess)
# Modify engine_path as required for your engine location & operating system
engine_path <- '//stockfish_13_win_x64_bmi2.exe'
engine <- uci_engine(engine_path)
pgn_path <- file.path(
  system.file(package = 'rbitr'),
  'extdata',
  'fools_mate.pgn'
)
evaluate_pgn(pgn_path, engine, n_pv = 1, depth = 1)
}
