% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_position.R
\name{evaluate_position}
\alias{evaluate_position}
\title{Evaluate a chess position}
\usage{
evaluate_position(
  position,
  engine_path,
  limiter,
  limit,
  n_cpus,
  n_pv,
  hash_size = NULL
)
}
\arguments{
\item{position}{A single-element character vector containing a series of
legal chess moves in long algebraic notation (LAN).}

\item{engine_path}{A single-element character vector of the path to a UCI
compatible chess engine.}

\item{limiter}{A single-element character vector indicating the desired
mode of search termination. Allowed values are 'depth' (to search a fixed
number of plies), 'nodes' (to search a fixed number of nodes), and
'movetime' (to search for a fixed number of milliseconds).}

\item{limit}{A single-element integer vector of the desired search depth
(# of plies), search nodes (# of nodes), or search time
(# of milliseconds).}

\item{n_cpus}{(Default = 1) A single-element integer vector of the number of
CPUs to use.}

\item{n_pv}{(Default = 1) A single-element integer vector of the desired
number of principal variations.}

\item{hash_size}{(Default = NULL) A single-element integer vector of the
desired hash size, in MB.}
}
\value{
A list containing a character vector of the engine's output.
}
\description{
Chess positions are analyzed using a
\href{http://wbec-ridderkerk.nl/html/UCIProtocol.html}{UCI compatible} chess
engine.
}
\examples{
position <- 'e2e4'
# Modify engine_path as required for your engine location & operating system
engine_path <- '//stockfish.exe'
evaluate_position(position, engine_path, n_pv = 1, n_cpus = 1,
                  limiter = 'depth', limit = 1)
}
\seealso{
The 'cram' functions condense analysis logs into data frames.
\itemize{
\item \code{\link[=cram_positionlog]{cram_positionlog()}} for condensing analysis of one position.
\item \code{\link[=cram_pgnlog]{cram_pgnlog()}} for condensing analysis of an entire pgn file.
}

The 'parse' functions extract specific data from analysis logs.
\itemize{
\item \code{\link[=parse_gamelog]{parse_gamelog()}} for extracting data from one evaluated game.
\item \code{\link[=parse_pgnlog]{parse_pgnlog()}} for extracting data from games in a pgn.
}

The 'evaluate' functions produce analysis logs.
\itemize{
\item \code{\link[=evaluate_game]{evaluate_game()}} for analyzing individual games.
\item \code{\link[=evaluate_pgn]{evaluate_pgn()}} for evaluating all the games in a PGN file.
}

Functions to load and save PGN files.
\itemize{
\item \code{\link[=get_pgn]{get_pgn()}} for loading PGN files.
\item \code{\link[=save_pgn]{save_pgn()}} for saving PGN files.
}
}
