% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eval.R
\name{extract_eval}
\alias{extract_eval}
\title{Extract Evaluations from Logs}
\usage{
extract_eval(evaluations, root)
}
\arguments{
\item{evaluations}{A recursive list of evaluation data.}

\item{root}{The type of evaluation to extract.}
}
\value{
The extracted evaluations.
}
\description{
Extract Evaluations from Logs
}
\examples{
evaluations <- list(list(list(scores = '146', pvs = 'd2d4')))
extract_eval(evaluations, 'scores')
extract_eval(evaluations, 'pvs')
}
