% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fen_to_board.R
\name{fen_row_to_vector}
\alias{fen_row_to_vector}
\title{Convert row from FEN string into 8x1 vector}
\usage{
fen_row_to_vector(row_string)
}
\arguments{
\item{row_string}{A string representing chessboard row in FEN format}
}
\value{
A vector representing a chessboard row, with one element per square
}
\description{
The function \code{fen_row_to_vector()} is a helper function for
\code{\link[=fen_to_board]{fen_to_board()}}.
}
