% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fen_to_board.R
\name{fen_to_board}
\alias{fen_to_board}
\title{Convert FEN notation to a chessboard}
\usage{
fen_to_board(fen = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1")
}
\arguments{
\item{fen}{A valid FEN string. Defaults to starting position if omitted.}
}
\value{
A named list containing data from the FEN string, including an 8x8
matrix of the board.
}
\description{
Forsyth-Edwards Notation (FEN) is a compact way to represent the complete
state of a chess game as a single line of ASCII text. In addition to showing
the position of the pieces on the board, it stores information about whose
turn it is, castling rights, en passant availability, the 50 move rule, and
the move number. A FEN string contains six fields separated by spaces.
}
\details{
The function \code{fen_to_board()} makes the information in a FEN string more
accessible for use in R. The board representation in converted into an 8x8
matrix and is stored in a named list along with the other fields. The names
are $board, $to_move, $castling_rights, $ep_target, $halfmove_clock, and
$fullmove_number. See
\href{http://www.saremba.de/chessgml/standards/pgn/pgn-complete.htm#c16.1}{section 16.1 of the PGN specification}
for detailed information about each field.

In FEN, the pieces are represented as letters, with lowercase for
black and uppercase for white.

p, P = pawn
n, N = knight
b, B = bishop
r, R = rook
q, Q = queen
k, K = king

Empty squares in FEN are represented by digits indicating a horizontal group
of that many empty squares in a row, with rows separated by the / character.
For example, the FEN string of the starting position is:

'rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1'

Valid FEN strings may be supplied to the \code{fen_to_board} function using the
\code{fen} argument. Currently, the function does not check the validity of the
FEN string, so it is up to the user to ensure that only valid FEN strings are
used. If no argument is supplied, the starting position is the default.

The FEN piece notation is preserved for matrix representation, but empty
squares are represented as empty strings rather than digits. Files a-h have
matrix row indices 1-8, and ranks 1-8 have matrix column indices 1-8.
Squares can be accessed by their coordinates as \code{board[rank, file]} (i.e.,
square b7 is at \code{board[7, 2]}). They may also be accessed by column names
(i.e., \code{board[7, 'b']}).
}
\note{
When printing a matrix in R, row 1 appears at the top. This is the
opposite convention from algebraic notation when the chessboard is viewed
with white at the bottom, where row 1 is the bottom row. Therefore
\code{print(fen_to_board()$board)} shows a board that has been flipped vertically,
as if viewed from below. Rotating the board does not fix the appearance,
since the columns will still appear to be in reverse order. To print the
board in the conventional manner, use the function \code{\link[=print_board]{print_board()}}.
}
\examples{
board <- fen_to_board()$board
# Show the piece on e1
board[1, 5]
# Show the entire board
print_board(board)
}
\seealso{
\itemize{
\item \code{\link[=board_to_fen]{board_to_fen()}} To convert a board matrix to a FEN string.
\item \code{\link[=print_board]{print_board()}} Prints the board with rows in the correct order.
}
}
