% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_big_pgn_tags.R
\name{get_big_pgn_tags}
\alias{get_big_pgn_tags}
\title{Get the tag names used in a large pgn.}
\usage{
get_big_pgn_tags(pgn_path, batch_size = 10^7, silent = FALSE)
}
\arguments{
\item{pgn_path}{A single-element character vector of the path to a pgn file.}

\item{batch_size}{(Default = 10^7) A single-element integer vector indicating
how many lines at a time are read from the file.}

\item{silent}{(Boolean, default = FALSE) Report progress?}
}
\value{
A character vector of the unique tag names found in the pgn.
}
\description{
This is a helper function for rbitr's \code{read_pgn()} to ensure that all
available tags will be read from the file. It is an alternative to rbitr's
\code{get_pgn_tags()} intended for large pgn files. Unlike \code{get_pgn_tags},
\code{get_big_pgn_tags()} only keeps a portion of the file in memory at a time,
and also periodically reports progress.
}
\details{
See the
\href{http://www.saremba.de/chessgml/standards/pgn/pgn-complete.htm#c8.1.1}{pgn specification}
for the definition of tag pairs.

Every game in the pgn will be searched, and the name of every tag
found will be reported, even if the tag only appears in only one game. If a
tag appears more than once in the file, it will only appear once in the
output; subsequent occurrences are ignored.
}
\examples{
pgn_path <- file.path(
  system.file(package = "rbitr"),
  'extdata',
  'fools_mate.pgn'
)
get_big_pgn_tags(pgn_path)
}
