% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clocks.R
\name{get_clocks}
\alias{get_clocks}
\title{Get clock data from pgn movetext}
\usage{
get_clocks(movetext)
}
\arguments{
\item{movetext}{A character vector of pgn movetext, where each vector entry
is for a separate game.}
}
\value{
A list containing numeric vectors of time remaining after each move,
in seconds. Each list entry is for a separate game.
}
\description{
Get the clock data (if present) from the movetext taken from a pgn file.
}
\details{
See the
\href{http://www.enpassant.dk/chess/palview/enhancedpgn.htm}{Portable Game Notation Specification and Implementation Guide (Supplement)}
for details on the clk command in pgn files.

The \code{get_clocks()} function is intended for use with data read from
a pgn using the \code{get_pgn()} function. A tibble obtained using \code{get_pgn()}
will have a $Movetext column from which the clock data can be extracted if
it is present.
}
\examples{
movetext <- c(
  '1. d4 {[\%clk 0:15:00]} d5 {[\%clk 0:15:00]} 2. Nc3 {[\%clk 0:14:48]}',
  '1. e4 {[\%clk 0:15:00]} e5 {[\%clk 0:15:00]} 2. Nf3 {[\%clk 0:14:38]}'
)
get_clocks(movetext)
}
