% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coercion.R
\name{get_coercion}
\alias{get_coercion}
\title{Gets the coercion of a chess position}
\usage{
get_coercion(positionlog)
}
\arguments{
\item{positionlog}{A character vector of engine analysis}
}
\value{
A single-element numeric vector of the \code{coercion} for the position.
}
\description{
The \code{get_coercion()} function calculates the difference in evaluation between
the best and second best moves in a chess position, referred to as the
\code{coercion} of the position.
}
\details{
In chess, forcing moves are those moves in which the opponent has
only a small number of choices that don't result in an unacceptable loss of
material or positional advantage. Guid and Bratko (see references) used the
difference in engine evaluation between the best and second best moves in a
position as a measure of how forcing a position is.

Given a log of engine analysis of a chess position produced by one
of \code{rbitr}'s evaluation functions, \code{get_coercion()} will compute the
difference in evaluation between the best and second best moves, referred
to here as the \code{coercion} of the position.

\code{get_coercion()} requires as input a \code{positionlog} of engine
analysis. The analysis must contain at least two principal variations
(\code{pv}'s) to allow the \code{coercion} to be determined. If only one PV exists,
a value of Inf is returned, or if no move is possible (i.e., in positions
of mate or stalemate) it will return \code{NA}. The highest depth of search
available will be used by default.
}
\examples{
positionlog <- c(
  "Stockfish 13 by the Stockfish developers (see AUTHORS file)",
  "readyok",
  "info string NNUE evaluation using nn-62ef826d1a6d.nnue enabled",
  "info depth 1 seldepth 1 multipv 1 score cp 161 nodes 59 nps 59000 tbhits 0 time 1 pv b1c3",
  "info depth 1 seldepth 1 multipv 2 score cp 146 nodes 59 nps 59000 tbhits 0 time 1 pv d2d4",
  "info depth 2 seldepth 2 multipv 1 score cp 195 nodes 147 nps 73500 tbhits 0 time 2 pv d2d4 g8f6",
  "info depth 2 seldepth 2 multipv 2 score cp 170 nodes 147 nps 73500 tbhits
  0 time 2 pv h2h4 g8f6 h4g5",
  "bestmove d2d4 ponder g8f6"
)
}
\references{
Guid, M.; Bratko, I. Computer Analysis of World Chess Champions. ICGA
Journal 2006, 29 (2), 65–73.
}
\seealso{
The 'evaluate' functions produce analysis logs.
\itemize{
\item \code{\link[=evaluate_position]{evaluate_position()}} for analyzing chess positions.
\item \code{\link[=evaluate_game]{evaluate_game()}} for analyzing individual games.
\item \code{\link[=evaluate_pgn]{evaluate_pgn()}} for evaluating all the games in a PGN file.
}
}
