% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cpl.R
\name{get_cpl}
\alias{get_cpl}
\title{Calculate centipawn loss.}
\usage{
get_cpl(movetext, engine, depth)
}
\arguments{
\item{movetext}{A single element character vector containing a sequence of
moves in SAN.}

\item{engine}{An engine handler generated by the
\href{https://github.com/rosawojciech/bigchess}{bigchess} function
\code{uci_engine()}.}

\item{depth}{A single-element integer vector of the desired search depth, in
plies.}
}
\value{
An integer vector of centipawn loss for each move.
}
\description{
Calculate centipawn loss.
}
