% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cpu_name.R
\name{get_cpu_name}
\alias{get_cpu_name}
\title{Get the name of the CPU}
\usage{
get_cpu_name()
}
\value{
A character string representing the name of the CPU.
}
\description{
This function attempts to return the name of the CPU model. See Note.
}
\details{
The function first detects the operating system using the
\code{Sys.info()} function in R. Depending on the operating system, it uses
OS-specific system commands to retrieve the CPU name:
\itemize{
\item Windows (PowerShell): \verb{(Get-WmiObject Win32_Processor).Name}
\item Linux: \verb{lscpu | grep "Model name:"}
\item macOS: \verb{sysctl -n machdep.cpu.brand_string}
}

For operating systems other than the above, it returns "Unknown CPU."
}
\note{
This function has only been tested in Windows 11. Users are advised to
manually test this function in their specific environment to ensure it
works as expected.
}
\examples{
get_cpu_name()
}
