% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_engine_match.R
\name{get_engine_match}
\alias{get_engine_match}
\title{Determine the ratio of a player's moves that match an engine's moves}
\usage{
get_engine_match(gamelog, movetext, depth = NULL)
}
\arguments{
\item{gamelog}{A list of engine output from rbitr's \code{evaluate_game()}
function.}

\item{movetext}{A single-element character vector containing a sequence of
moves in standard algebraic notation (SAN).}

\item{depth}{(Optional, default = NULL) A single-element integer vector
indicating which search depth to return. The value must not be less than 1
or greater than the maximum depth reported by the engine. A value of NULL
returns data for the maximum available depth.}
}
\value{
A list containing the ratio of engine-matching moves to the total
number of moves for both $white_ratio and $black_ratio.
}
\description{
The \code{get_engine_match()} function calculates the fraction of moves played
that correspond to the 'best' move as determined by a chess engine.
}
\details{
The \code{get_engine_match()} function requires a gamelog in the same
format as the one returned by \code{evaluate_game()}. It also requires the
moves of the game to be supplied in the same format provided by \code{get_pgn()}.

If desired, the analysis depth used to determine the
best move may be specified using the \code{depth} parameter. If \code{depth} is set
to a value higher than what is available for some moves, \code{evaluate_game()}
will return an error. If no value for \code{depth} is provided, the deepest
analysis at each ply will be used.

The ratio is determined by counting the number of moves that are the
same as the engine's preferred moves, and then dividing by the total number
of moves in the game:
bestmove_ratio = # of moves matching the engine / total # of moves
}
\note{
By itself, observing a player to have a high percentage of moves that
are the same as engine moves is \emph{not} necessarily evidence of illicit
computer assistance. Having a high percentage can be a consequence of a
game having many \emph{forcing} lines. It is also the case that if many games
are analyzed, it becomes likely that some of them will have high matching
percentages by chance alone. Prof. Ken Regan calls this
\href{https://cse.buffalo.edu/~regan/chess/fidelity/Golfers.html}{the parable of the golfers}.
}
\examples{
movetext <- '1. e4 g5 2. Nc3 f5 3. Qh5# 1-0'
# Modify engine_path as required for your engine location & operating system
engine_path <- '//stockfish.exe'
gamelog <- evaluate_game(movetext, engine_path, n_pv = 1, limiter = 'depth',
                         limit = 1)
get_engine_match(gamelog, movetext)
}
\seealso{
\itemize{
\item \code{\link[=get_pgn]{get_pgn()}} for loading the movetext from a pgn file.
\item \code{\link[=evaluate_game]{evaluate_game()}} for creating a gamelog.
}
}
