% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imb.R
\name{get_imb}
\alias{get_imb}
\title{Identify inaccuracies, mistakes, and blunders.}
\usage{
get_imb(
  scores,
  moves,
  bestmoves,
  color,
  cap = 1000,
  cap_action = "replace",
  first_ply = 1,
  to_move = "white",
  mate = 50000
)
}
\arguments{
\item{scores}{A numeric or integer vector of chess engine evaluations.}

\item{moves}{A character vector of moves in the same format as \code{bestmoves}.}

\item{bestmoves}{A character vector of moves in the same format as \code{moves}.}

\item{color}{A single-element character vector indicating which side to
analyze. Allowed values are 'black' or 'white'.}

\item{cap}{(Default = 1000) A single-element integer or numeric vector of the
maximum allowed value for centipawn score. The default of 1000 centipawns
is the value used by lichess.}

\item{cap_action}{(Default = 'replace') A single-element character vector
indicating whether to exclude scores outside the range [-cap, cap], or to
replace them with the cap. Allowed values are 'exclude', 'replace', or
'none'. The default of 'replace' is required to mimic lichess.}

\item{first_ply}{(Default = 1) A single-element integer vector indicating the
first ply to be included in the calculation. May not be larger than the
number of elements in \code{score}. A value of 1 is required to mimic lichess.}

\item{to_move}{(Default = 'white') A single-element character vector
indicating which side's turn it is. The default value of 'white' is
required to mimic lichess.}

\item{mate}{(Default = 50000) A single-element integer or numeric vector of
the centipawn value to assign for mate. Should always be set higher than
the setting for cap. It should also be higher than the highest numerical
evaluation in the game to avoid unexpected behaviour.}
}
\value{
A named list of $inaccuracies, $mistakes, and $blunders.
}
\description{
The function \code{get_imb()} is meant to mimic the inaccuracy, mistake, and
blunder stats provided by the analysis tool on
\href{http://lichess.org}{lichess.org}. When analyzing using the same engine and
the same engine settings that lichess uses, \code{get_imb()} should produce
numbers very close to those given by lichess. Some differences are expected
however, since most chess engines are non-deterministic for a variety of
reasons.
}
\details{
The user must provide a vector of evaluations for the game
(\code{scores}). These should include an evaluation for the initial position,
before white has moved. In addition, a vector of the game's moves
(\code{moves}), and a vector of the engine's preferred moves (\code{bestmoves}) must
also be provided. The move data is used to avoid marking the best available
move as an inaccuracy, mistake, or blunder. The \code{color} parameter is used
to indicate which side to evaluate.
}
\note{
The engine used by lichess.org changes from time to time as engine
technology improves. Refer to their website for information on the engine
currently in use.
}
\examples{
scores <- c(12, -171, -72, -50000, -50000)
moves <- c("g2g4", "e7e6", "f2f4", "d8h4")
bestmoves <- c("d2d4", "d7d5", "g1f3", "d8h4")
get_imb(scores, moves, bestmoves, color = 'white')
}
