% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last_move.R
\name{get_last_move}
\alias{get_last_move}
\title{Get the last move.}
\usage{
get_last_move(current_position)
}
\arguments{
\item{current_position}{A single element character vector containing a legal
sequence of chess moves in long algebraic notation.}
}
\value{
A single-element character vector of the most recently played move in
LAN format.
}
\description{
Given a chess position represented as a LAN format sequence of moves, return
the most recent move.
}
\examples{
movetext <- 'e2e4 g7g5 b1c3 f7f5 d1h5'
get_last_move(movetext)
}
