% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_legal_moves.R
\name{get_legal_moves}
\alias{get_legal_moves}
\title{Get all legal moves for the provided chess position.}
\usage{
get_legal_moves(position, engine_path, n_pv = 256L)
}
\arguments{
\item{position}{A single-element character vector containing a sequence of
moves in long algebraic notation (LAN).}

\item{engine_path}{A single-element character vector of the path to a UCI
compatible chess engine.}

\item{n_pv}{(Default = 256L) A single-element integer vector used to set the
number of principal variations for the engine to consider.}
}
\value{
A character vector containing all of the legal moves in a position in
long algebraic notation (LAN).
}
\description{
Generates a list of all legal moves from the provided chess position.
}
\details{
The function \code{get_legal_moves()} passes the position to an external
\href{http://wbec-ridderkerk.nl/html/UCIProtocol.html}{UCI compatible} chess
engine, which must be supplied by the user. The path to the engine's
executable file is given by the engine_path parameter. The engine is asked
to return an analysis with minimum depth, and maximum number of principal
variations (PVs). The legal moves are obtained by parsing the PV output. If
the number of PVs given by the engine is limited, then it is possible that
not every possible legal move will be generated.

The position to be analyzed is represented as a sequence of legal
chess moves. As a consequence, it is not possible to generate moves for
positions that can not be arrived at by a sequence of legal moves.
}
\note{
As of version 18, stockfish has a limit of 256 PVs. The maximum number
of legal moves in a legal chess position has been \href{https://lichess.org/@/Tobs40/blog/why-a-reachable-position-can-have-at-most-218-playable-moves/a5xdxeqs}{found to be 218.}
The n_pv parameter sets the upper limit for the number of moves that can be
detected. Setting this value to less than the number of legal moves will
cause the function to incorrectly count the number of moves. It is provided
as an adjustable parameter in case future versions of stockfish or other
engines require lower numbers to function. In general, it is recommended to
just use the default value.
}
\examples{
# Show all possible white first moves
# Modify engine_path as required for your engine location & operating system
engine_path <- '//stockfish.exe'
get_legal_moves('', engine_path)
}
