% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_move_times.R
\name{get_move_times}
\alias{get_move_times}
\title{Convert clock data to move times}
\usage{
get_move_times(clock, increment, color)
}
\arguments{
\item{clock}{An integer or numeric vector of chess clock times at the end of
each half move (in seconds).}

\item{increment}{A single-element integer or numeric vector of the increment
time (in seconds).}

\item{color}{A character vector indicating which side to calculate (either
black or white).}
}
\value{
A numeric vector of the move times (if present) in seconds. The
vector starts at the second move (the first move is zero seconds).
}
\description{
Clock data (in seconds) is converted into move times (in seconds).
}
\details{
Clock times and increments are typically obtained using rbitr's
\code{get_clocks()}, and \code{get_increments()} functions. The move times start from
the second move (the first move is taken to be zero seconds).
}
\note{
The author has noted that clock data in pgns from lichess.org may
sometimes give negative move times. The author speculates that these errors
may be due to unrecorded additions of time during the game. Since lichess has
options to allow moves to be taken back, or for the opponent to grant more
time during the game, it may be possible that these actions are not being
accurately reflected in the clock times recorded in the pgn. If unrecorded
additions of time were to occur in a given move, then the calculated move
time will be shorter than the actual move time. Such an error would only be
detected if it resulted in a negative move time. Whatever the cause, in the
rare cases where negative move times are encountered, the negative value is
replaced with NA.
}
\examples{
clock <- c(900, 900, 888, 878, 878, 858)
get_move_times(clock, 8, 'white')
get_move_times(clock, 8, 'black')
}
