% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgn_data_manipulators.R
\name{get_moves}
\alias{get_moves}
\title{Convert movetext to individual moves.}
\usage{
get_moves(movetext)
}
\arguments{
\item{movetext}{A character vector containing a sequence of moves in SAN.}
}
\value{
A recursive list of moves.
}
\description{
Convert movetext in standard algebraic notation (SAN) to individual moves in
long algebraic notation (LAN).
}
\examples{
movetext <- '1. e4 g5 2. Nc3 f5 3. Qh5#'
get_moves(movetext)
}
\seealso{
Other pgn data manipulators: 
\code{\link{calculate_move_times}()},
\code{\link{get_clocks}()},
\code{\link{get_increments}()}
}
\concept{pgn data manipulators}
