% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgn_file_management_tools.R
\name{get_pgn_tags}
\alias{get_pgn_tags}
\title{Get the tag names used in a pgn}
\usage{
get_pgn_tags(pgn_path)
}
\arguments{
\item{pgn_path}{A single-element character vector of the path to a pgn file.}
}
\value{
A character vector of the unique tag names found in the pgn.
}
\description{
This is a helper function for \code{rbitr::read_pgn} to ensure that all available
tags  will be read from the file.
}
\details{
See the pgn specification for the definition of tag pairs
(http://www.saremba.de/chessgml/standards/pgn/pgn-complete.htm#c8.1.1).

Every game in the pgn will be searched, and the name of every tag
found will be reported, even if the tag only appears in only one game. If a
tag appears more than once in the file, it will only appear once in the
output; subsequent occurrences are ignored.
}
\examples{
pgn_path <- file.path(
  system.file(package = "rbitr"),
  'extdata',
  'fools_mate.pgn'
)
get_pgn_tags(pgn_path)
}
\seealso{
Other pgn file management tools: 
\code{\link{read_pgn}()}
}
\concept{pgn file management tools}
