% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_points.R
\name{get_points}
\alias{get_points}
\title{Get points from chess match results}
\usage{
get_points(results)
}
\arguments{
\item{results}{A vector of strings representing the results of chess matches.}
}
\value{
A numeric vector representing the points from the chess matches.
}
\description{
This function takes a vector of chess match results as input and
converts them to tournament points.
}
\details{
Each result should be one of the following strings: "0-1", "1-0",
"1/2-1/2", or "*". The function returns 0, 1, 0.5, or NA respectively for
each result. If a result is not one of these four strings, it throws an
error.
}
\examples{
get_points(c("0-1", "1-0", "1/2-1/2", "*")) # returns c(0, 1, 0.5, NA)
}
