% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_position.R
\name{get_position}
\alias{get_position}
\title{Generate a position from movetext}
\usage{
get_position(movetext, ply)
}
\arguments{
\item{movetext}{A string containing chess moves in SAN.}

\item{ply}{An integer specifying a number of ply.}
}
\value{
A character vector of chess positions.
}
\description{
This function accepts a string containing chess moves in Standard Algebraic
Notation (SAN), and an integer specifying a number of ply. It returns a
shorter string of chess moves in Long Algebraic Notation (LAN).
}
\details{
Positions are represented by the series of chess moves that led to
the position. The \code{ply} parameter specifies which position should be
represented in LAN.

This function is useful for generating positions to be analyzed by a chess
engine. The initial position is represented by an empty string "".

The function assumes the input to be valid SAN conforming to the PGN
specification. It does not validate the input or check for legal chess
moves.
}
\examples{
movetext <- "1. e4 e5 2. Nf3 Nc6 3. Bb5 a6"
get_position(movetext, 3)

}
