% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgn_data_manipulators.R
\name{get_positions}
\alias{get_positions}
\title{Given a movetext string, generate a vector of chess positions}
\usage{
get_positions(movetext)
}
\arguments{
\item{movetext}{A single element character vector containing a sequence of
moves in SAN}
}
\value{
A character vector where each element represents a successive
position in the game as a sequence of moves in LAN
}
\description{
Given a single-element character vector containing a series of chess \emph{moves}
(typically from a pgn) in standard algebraic notation (SAN), return a
character vector containing the game \emph{positions} in long algebraic notation
(LAN), to enable evaluation by a chess engine. The positions are represented
as sequences of moves.
}
