% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_previous_position.R
\name{get_previous_position}
\alias{get_previous_position}
\title{Get the previous position.}
\usage{
get_previous_position(current_position)
}
\arguments{
\item{current_position}{A single element character vector of moves in}
}
\value{
A single-element character vector containing a string of moves in LAN
}
\description{
Trim the last ply from a position string to generate the previous position.
}
\examples{
movetext <- 'e2e4 g7g5 b1c3 f7f5 d1h5'
get_previous_position(movetext)
}
