% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_result.R
\name{get_result}
\alias{get_result}
\title{Get the result of a chess game from SAN format moves}
\usage{
get_result(movetext)
}
\arguments{
\item{movetext}{A character string of SAN format chess moves.}
}
\value{
The termination marker of the game if found, otherwise NA.
}
\description{
This function takes a character string of SAN format chess moves and returns
the result of the game.
}
\details{
The functions looks for one of the four possible game termination
markers defined by the PGN specification ("0-1", "1-0", "1/2-1/2", or "*")
and returns the marker if found. If no marker is found it returns NA.
}
\note{
If multiple markers are present, only the first one is returned.
}
\examples{
get_result("1.e4 e5 2.Nf3 Nc6 3.Bb5 a6 4.Ba4 Nf6 1-0")
get_result("1.e4 c5 2.Nf3 d6 3.d4 cxd4 4.Nxd4 Nf6 1/2-1/2")
}
