% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_columns.R
\name{harmonize_columns}
\alias{harmonize_columns}
\title{Harmonize columns in a list of data frames}
\usage{
harmonize_columns(df_list)
}
\arguments{
\item{df_list}{A list of data frames}
}
\value{
A list of data frames with harmonized columns
}
\description{
This function takes a list of data frames and adds missing columns to each
data frame so that all data frames have the same columns in the same order.
Any added columns will be populated with NAs. The function can be used to
prepare data frames for an rbind when the columns are not identical.
}
