% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_check.R
\name{is_check}
\alias{is_check}
\title{Look for checks from enemy pieces}
\usage{
is_check(board, color)
}
\arguments{
\item{board}{An 8x8 matrix representing a chess position. See
\code{\link[=fen_to_board]{fen_to_board()}} for details of the board representation.}

\item{color}{A string ("white" or "black"), indicating which king to examine.}
}
\value{
TRUE if the king is in check from an enemy piece, FALSE otherwise.
}
\description{
This function returns TRUE if there are checks from enemy pieces, and FALSE
otherwise.
}
\examples{
board <- matrix(rep('', 64), nrow = 8)
board[5, 5] <- 'k'
print(board)
is_check(board, 'black')
board[4, 4] <- 'P'
print(board)
is_check(board, 'black')
}
