% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_insufficient.R
\name{is_insufficient}
\alias{is_insufficient}
\title{Check for draw by insufficient material}
\usage{
is_insufficient(board)
}
\arguments{
\item{board}{An 8x8 matrix representing the chess board.}
}
\value{
A Boolean indicating whether there is insufficient material.
}
\description{
If mate is impossible due to lack of material, \code{is_insufficient()} returns
\code{TRUE}. In cases where mate is theoretically possible \code{is_insufficient()}
will return \code{FALSE} even if mate can't be forced.
}
\details{
Only the following material combinations are considered to be insufficient:
\itemize{
\item King versus king
\item King and bishop versus king
\item King and knight versus king
\item King versus king with any number of bishops on either side, if all bishops
are on the same color squares
}
}
\examples{
board <- matrix(rep("", 64), nrow = 8)
board[c(1, 64)] <- c("K", "k") # King versus king
is_insufficient(board)
}
