% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_san.R
\name{is_san}
\alias{is_san}
\title{Check if a string contains chess moves in standard algebraic notation (SAN)}
\usage{
is_san(movetext)
}
\arguments{
\item{movetext}{A character string, possibly containing a series of chess
moves in SAN.}
}
\value{
TRUE if the string contains chess moves in SAN format, FALSE
otherwise.
}
\description{
The function uses regular expressions to check if the string's tokens match
the expected pattern for chess moves in SAN.
}
\details{
Before making the check, an attempt is made to remove annotations,
game termination markers, or other content that would disrupt the regular
expression pattern. Note that this function does not check if the moves are
legal.
}
\examples{
is_san("1. e4! {best by test} e5 2. Nf3 {black resigns in shame} 1-0")
is_san("e2e4 e7e5 g1f3")

}
