% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cram_positionlog.R
\name{is_uci_move}
\alias{is_uci_move}
\title{Check if a move is in UCI format}
\usage{
is_uci_move(move)
}
\arguments{
\item{move}{A single-element character string representing a chess move.}
}
\value{
A logical value indicating if the move is in UCI format.
}
\description{
The \code{is_uci_move()} function checks to see if a given move is in the
\href{https://github.com/fsmosca/UCIChessEngineProtocol}{UCI} format. It is an
internal helper function optimized for speed. Input validation should
therefore be handled by a parent function.
}
\details{
A valid UCI move is 4 or 5 characters long, with the first and third
characters being letters from 'a' to 'h', the second and fourth characters
being numbers from 1 to 8, and an optional fifth character indicating a
promoted piece ('q', 'r', 'b', or 'n').
}
\note{
The \code{is_uci_move()} function does \emph{not} check to see if the move is a
legal chess move. It only checks to see of the move consists of two valid
chessboard squares and an optional valid promotion piece.
}
