% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_ucilan.R
\name{is_ucilan}
\alias{is_ucilan}
\title{Check if a string contains chess moves in UCI-format long algebraic notation
(LAN)}
\usage{
is_ucilan(movetext)
}
\arguments{
\item{movetext}{A character string, possibly containing a series of chess
moves in SAN.}
}
\value{
TRUE if the string contains valid chess moves in LAN, FALSE otherwise
}
\description{
The function uses regular expressions to check if the string's tokens match
the expected pattern for chess moves in LAN.
}
\details{
UCI-format long algebraic notation is described in the \href{http://wbec-ridderkerk.nl/html/UCIProtocol.html}{UCI protocol}. Note that this
function does not check if the moves are legal.
}
\examples{
is_ucilan("e2e4 e7e5 g1f3 b8c6")
is_ucilan("1. e4 e5 2. Nf3 Nc6")

}
