% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/can_king_move.R
\name{king_can_move}
\alias{king_can_move}
\alias{can_king_move}
\title{Check if the king can move}
\usage{
king_can_move(board, color = board$to_move)

can_king_move(board)
}
\arguments{
\item{board}{A list containing the current state of the chess board.}

\item{color}{An optional character vector giving the color of the king.
Defaults to the color of the player whose turn it is to move.}
}
\value{
A logical value indicating whether the king can move. Returns TRUE if
the king has at least one valid move, and FALSE otherwise.
}
\description{
This function checks if the king has any legal moves.
}
\details{
See \code{\link[=fen_to_board]{fen_to_board()}} for details of the chess board format.
This function is used by other functions when testing for checkmate or
stalemate. \code{can_king_move()} is an alias for \code{king_can_move()}.
}
\examples{
board <- fen_to_board() # Starting position
king_can_move(board)

}
