% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lan_to_boards.R
\name{lan_to_boards}
\alias{lan_to_boards}
\title{Convert UCI LAN to a list of boards}
\usage{
lan_to_boards(lan)
}
\arguments{
\item{lan}{A string of chess moves in UCI LAN.}
}
\value{
A list of boards representing the state of the game after each move.
}
\description{
This function takes a string of moves in UCI LAN (Long Algebraic Notation)
and converts it into a list of boards. Each board represents the state of the
game after each move.
}
\examples{
lan_to_boards("e2e4 e7e5 g1f3 b8c6 f1b5")
}
