% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lan_to_coordinates.R
\name{lan_to_coordinates}
\alias{lan_to_coordinates}
\title{Convert a chess move in long algebraic notation to coordinates}
\usage{
lan_to_coordinates(lan)
}
\arguments{
\item{lan}{A string representing a chess move in
\href{http://wbec-ridderkerk.nl/html/UCIProtocol.html}{UCI} long algebraic
notation (e.g., "e2e4", "e7e8q").}
}
\value{
A list containing the original coordinates, destination coordinates,
and the promoted piece (if any).
}
\description{
Convert a chess move in long algebraic notation to coordinates
}
\examples{
lan_to_coordinates("e2e4")
lan_to_coordinates("e7e8q")
}
