% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lan_to_fen.R
\name{lan_to_fen}
\alias{lan_to_fen}
\title{Convert a series of moves in long algebraic notation to a FEN string}
\usage{
lan_to_fen(lan)
}
\arguments{
\item{lan}{A string representing a series of chess moves in
UCI long algebraic notation (e.g., "e2e4 e7e5 g1f3").}
}
\value{
A string representing the final position after the series of moves
in FEN format.
}
\description{
Convert a series of moves in long algebraic notation to a FEN string
}
\examples{
lan_to_fen("e2e4 e7e5 g1f3")
}
