% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lichess_advantage_plot.R
\name{lichess_advantage_plot}
\alias{lichess_advantage_plot}
\title{Lichess advantage plot}
\usage{
lichess_advantage_plot(scores)
}
\arguments{
\item{scores}{A numeric or integer vector of positional evaluations (in
centipawns.)}
}
\value{
A ggplot object of the plotted data.
}
\description{
Generate an advantage plot similar to the ones in the "Computer analysis" tab
on lichess.org.
}
\details{
Given a vector of positional evaluations (in centipawns), the
results are scaled using the same exponential function used by lichess.org,
and the results are plotted using the same color scheme.
}
\examples{
  scores <- c(15, 5, 29, -94, 67, 76, 154, -31, 1000, 1000)
  lichess_advantage_plot(scores)
}
