% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lichess_plot.R
\name{lichess_plot}
\alias{lichess_plot}
\title{Lichess plot}
\usage{
lichess_plot(
  pgn_path,
  game_number,
  engine_path,
  n_cpus = 1,
  use_pgn_evals = TRUE
)
}
\arguments{
\item{pgn_path}{A single-element character vector of the path to a pgn file.}

\item{game_number}{A single-element integer vector indicating which game in
the pgn to plot.}

\item{engine_path}{A single-element character vector of the path to a UCI
compatible chess engine.}

\item{n_cpus}{(Default = 1) A single-element integer vector indicating how
many cpus to use for analysis.}

\item{use_pgn_evals}{(Default = TRUE) A single-element boolean indicating
whether to use evals from the pgn, if present.}
}
\value{
A ggplot object of the plotted data.
}
\description{
Generate a plot of game statistics in the style of
\href{http://lichess.org}{lichess.org}.
}
\details{
The function\code{lichess_plot()} will first look for existing analysis in the pgn
file, and if not found, it will also look for saved analysis. If no
analysis data is found, it will analyze the game to a depth of 2250000
nodes.
}
\examples{
pgn_path <- file.path(
  system.file(package = 'rbitr'),
  'extdata',
  'short_game.pgn'
)
# Modify engine_path as required for your engine location & operating system
engine_path <- '//stockfish_13_win_x64_bmi2.exe'
lichess_plot(pgn_path, game_number = 1, engine_path)
}
