% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbitr_helpers.R
\name{make_table}
\alias{make_table}
\title{Make a table}
\usage{
make_table(col_a, col_b)
}
\arguments{
\item{col_a}{A character vector of the rows of column A.}

\item{col_b}{A character vector of the rows of column B.}
}
\value{
A single-element character vector of text that will be a two-column
table when rendered with a monospaced font.
}
\description{
Makes a two-column table that will be justified when rendered with a
monospaced font.
}
\details{
A character string will be generated that produces a two-column
table when rendered in a monospaced font. The first column will be right-
justified, and the second column will be left-justified. Rows will be
separated by newline characters ('\\n').
}
\examples{
cat(make_table(c('1', '2', '11'), c('one', 'two', 'eleven')))
}
