% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_exposes_king.R
\name{move_exposes_king}
\alias{move_exposes_king}
\title{Does the move expose the same-color king to check?}
\usage{
move_exposes_king(board, coordinates, new_coordinates, color)
}
\arguments{
\item{board}{A list containing the current state of the chess board.}

\item{coordinates}{A vector of length 2 representing the coordinates of the
piece on the board before the move. The first element is the row number and
the second element is the column number.}

\item{new_coordinates}{A vector of length 2 representing the coordinates of
the piece on the board after the move. The first element is the row number
and the second element is the column number.}

\item{color}{A character vector giving the color of the piece.}
}
\value{
A Boolean indicating whether the move results in check.
}
\description{
This function creates a copy of the board and simulates a move. It then
checks if the same-color king would be in check after the move.
}
\examples{
board <- fen_to_board()
move_exposes_king(board, c(2, 1), c(3, 1), 'white')

}
