% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbitr_helpers.R
\name{nested_lapply}
\alias{nested_lapply}
\title{Apply \code{input_function} to the root level elements of a nested list.}
\usage{
nested_lapply(x, input_function, ...)
}
\arguments{
\item{x}{A nested list.}

\item{input_function}{A function that takes the root level elements of x as}

\item{...}{Further arguments to \code{input_function}.}
}
\value{
A list with the same structure as x, where the root level elements
have been replaced with the outputs of \code{input_function}.
}
\description{
Apply \code{input_function} to the root level elements of a nested list.
}
