% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_eval_logs.R
\name{parse_eval_logs}
\alias{parse_eval_logs}
\title{Parse Evaluation Logs}
\usage{
parse_eval_logs(eval_logs, depth = NULL)
}
\arguments{
\item{eval_logs}{A list of ucilogs.}

\item{depth}{(Optional, default = NULL) A single-element integer vector
indicating which search depth to return. The value must not be less than 1
or greater than the maximum depth reported by the engine. A value of NULL
returns data for the maximum depth.}
}
\value{
A recursive list of ucilogs.
}
\description{
Parse Evaluation Logs
}
\examples{
eval_logs <- list(list(c(
  "Stockfish 13 by the Stockfish developers (see AUTHORS file)",
  "readyok",
  "info string NNUE evaluation using nn-62ef826d1a6d.nnue enabled",
  "info depth 1 seldepth 1 multipv 1 score cp 146 nodes 30 nps 30000 tbhits 0 time 1 pv d2d4",
  "bestmove d2d4"
)))
parse_eval_logs(eval_logs)
}
