% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gamelog.R
\name{parse_gamelog}
\alias{parse_gamelog}
\title{Extract data from the output of rbitr's \code{evaluate_game()} function.}
\usage{
parse_gamelog(gamelog, target, depth = NULL)
}
\arguments{
\item{gamelog}{A list of engine output from rbitr's \code{evaluate_game()}
function.}

\item{target}{A single-element character vector of the output to return.
Allowed values are 'score' for the evaluation in centipawns, 'pv' for the
principal variation, or 'bestmove' for the best move.}

\item{depth}{(Optional, default = NULL) A single-element integer vector
indicating which search depth to return. The value must not be less than 1
or greater than the maximum depth reported by the engine. A value of NULL
returns data for the maximum depth.}
}
\value{
A vector of the extracted data.
}
\description{
The output of rbitr's \code{evaluate_game()} function consists of a list of UCI
engine output, where each entry in the list is for a single position. The
evaluations and variations are embedded within a lot of other data. The
\code{parse_gamelog()} function can extract positional evaluations in centipawns
(score), principal variations (pv), or best moves (bestmove).
}
\details{
The data for the deepest available search depth is returned by
default. See
\href{http://wbec-ridderkerk.nl/html/UCIProtocol.html}{the UCI protocol} for
details of UCI engine output.
}
\examples{
gamelog <- list(c(
  "Stockfish 13 by the Stockfish developers (see AUTHORS file)",
  "readyok",
  "info string NNUE evaluation using nn-62ef826d1a6d.nnue enabled",
  "info depth 1 seldepth 1 multipv 1 score cp 146 nodes 30 nps 30000 tbhits 0 time 1 pv d2d4",
  "bestmove d2d4"
))
parse_gamelog(gamelog, 'score')
parse_gamelog(gamelog, 'pv')
parse_gamelog(gamelog, 'bestmove')
}
