% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gamelog.R
\name{parse_gamelog_bestmove}
\alias{parse_gamelog_bestmove}
\title{Extract the best moves from rbitr's \code{evaluate_game()} output}
\usage{
parse_gamelog_bestmove(gamelog)
}
\arguments{
\item{gamelog}{A list of engine output from rbitr's \code{evaluate_game()}
function.}
}
\value{
A list of character vectors of the extracted best moves for each
position.
}
\description{
\code{parse_gamelog_bestmove()} is an internal function called by
\code{parse_gamelog()}.
}
\details{
Returns the engine's calculated best move (bestmove) for each
position in the game.  See the documentation for \code{parse_gamelog()} for
full details.
}
\seealso{
\itemize{
\item \code{\link[=parse_gamelog]{parse_gamelog()}}
}
}
