% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_movetext.R
\name{parse_movetext}
\alias{parse_movetext}
\title{Parse movetext}
\usage{
parse_movetext(movetext, cmd_name)
}
\arguments{
\item{movetext}{A character vector of pgn movetext, where each vector entry
is for a separate game.}

\item{cmd_name}{A single-element character vector of the command to parse}
}
\value{
A list containing numeric vectors of the specified tag value. Each
list entry is for a separate game. For \code{get_clocks()}, the value will be
converted to seconds.
}
\description{
The function \code{parse_movetext()} is the parser behind \code{get_clocks()} and
\code{get_evals()}.
}
