% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cram_positionlog.R
\name{parse_split_engine_line}
\alias{parse_split_engine_line}
\title{Helper function for extracting data from UCI chess engine output}
\usage{
parse_split_engine_line(tag, split_engine_line)
}
\arguments{
\item{tag}{A single-element character vector of a tag name that appears in
the engine output.}

\item{split_engine_line}{A character vector generated by using \code{strsplit()}
on a line of UCI engine output.}
}
\value{
A single-element character vector of the value that follows \code{tag} in
the output of a UCI chess engine.
}
\description{
Helper function for extracting data from UCI chess engine output
}
\details{
The \code{parse_split_engine_line()} function is optimized for speed for
large data sets. Input validation should therefore be handled by a parent
function.

\code{parse_split_engine_line()} expects a single line of UCI engine
output that has been split using \code{strsplit()}, with \code{split} = ' '. It will
return the value immediately following the specified \code{tag}. If the tag is
not present, it will return NA.

The \href{https://github.com/fsmosca/UCIChessEngineProtocol}{UCI}
protocol lists a number of tags that may appear in engine output. Most tags
have values that are either numeric, or that are a series of chess moves in
long algebraic notation (LAN):

Numeric tags: depth, seldepth, multipv, time, nodes, currmovenumber,
hashfull, nps, tbhits, sbhits, cpuload
Move tags: bestmove, ponder, pv, currmove, refutation
Special tags: score, string, currline

In the case of move tags, \code{parse_split_engine_line()} will check for the
next move until it reaches the end of the series, and return the moves as a
string separated by spaces.

The tags \code{score}, \code{string}, and \code{currline} differ from the rest. The
\code{score} tag has 4 different types of value (\if{html}{\out{<x>}} indicates numeric):
cp: A score (centipawns), 'cp \if{html}{\out{<x>}}'
mate: Moves until mate, 'mate \if{html}{\out{<x>}}'
lowerbound: A lower bound for the score (centipawns), '\if{html}{\out{<x>}} lowerbound'
upperbound: An upper bound for the score (centipawns), '\if{html}{\out{<x>}} upperbound'

The \code{string} tag can have any value, and it runs to the end of the line.

The \code{currline} has a slightly different format than the other move tags.
The moves may be preceded by a number indicating which cpu generated the
moves. In the case of only 1 cpu, the number may be omitted.

Some engines may use tags not listed in the UCI protocol. The
\code{parse_split_engine_line()} function may be able to deal with such tags if
they conform to the general format of other UCI tags.
}
