% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_tools.R
\name{parse_ucilog}
\alias{parse_ucilog}
\title{Extract principal variations (PVs) and scores from UCI chess engine output.}
\usage{
parse_ucilog(ucilog, depth = NULL)
}
\arguments{
\item{ucilog}{A character vector of engine output.}

\item{depth}{(Optional, default = NULL) A single-element integer vector
indicating which search depth to return. The value must not be less than 1
or greater than the maximum depth reported by the engine. A value of NULL
returns data for the maximum depth}
}
\value{
A recursive list containing the extracted scores and PVs.
}
\description{
Parse the output of a uci-compatible chess engine for the principal
variations and scores.
}
\details{
The data for the deepest available search depth is returned by
default. See
\href{http://wbec-ridderkerk.nl/html/UCIProtocol.html}{the UCI protocol} for
details of UCI engine output.
}
\examples{
ucilog <- list(c(
  "Stockfish 13 by the Stockfish developers (see AUTHORS file)",
  "readyok",
  "readyok",
  "readyok",
  "readyok",
  "info string NNUE evaluation using nn-62ef826d1a6d.nnue enabled",
  "info depth 1 seldepth 1 multipv 1 score cp 146 nodes 30 nps 30000 tbhits 0 time 1 pv d2d4",
  "bestmove d2d4"
))
parse_ucilog(ucilog)
}
