% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_board.R
\name{print_board}
\alias{print_board}
\title{Print a chessboard in the conventional orientation}
\usage{
print_board(board, pov = "white")
}
\arguments{
\item{board}{An 8x8 matrix representing a chessboard. See
\code{\link[=fen_to_board]{fen_to_board()}} for details of the board representation.}

\item{pov}{(Default = 'white') A string indicating which side's point of view
should be shown. Allowed values are 'black' and 'white'.}
}
\value{
Prints a chessboard in the correct orientation.
}
\description{
Prints a chessboard with the rows arranged bottom to top.
}
\details{
The conventional orientation of a 2-D chessboard places row 1 at
the bottom from white's point of view. R's \code{print()} function prints matrices
with row 1 at the top, causing chessboard rows to appear in reverse order,
resulting in a mirror image of the board. The \code{print_board()} function simply
a wrapper for \code{print()} that modifies the board before printing to give a
true image of the board.
}
\examples{
board <- fen_to_board()$board
print_board(board)
}
