% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgn_file_management_tools.R
\name{read_pgn}
\alias{read_pgn}
\title{Read a pgn file}
\usage{
read_pgn(pgn_path, verbose = FALSE)
}
\arguments{
\item{pgn_path}{A single-element character vector of the path to a pgn file.}

\item{verbose}{(Boolean, default = TRUE) Print a message showing the file
being read. Useful to show progress when batch processing a lot of pgn
files.}
}
\value{
A data frame of the pgn data (see \code{bigchess::read.pgn} for further
details)
}
\description{
Convenience wrapper specifying preferred rbitr defaults for
\code{bigchess::read.pgn}. The contents of the pgn file are read into a data
frame with  the tag names as column names. All tags in the pgn will be read.
}
\examples{
pgn_path <- file.path(
  system.file(package = 'rbitr'),
  'extdata',
  'fools_mate.pgn'
)
read_pgn(pgn_path)
}
\seealso{
Other pgn file management tools: 
\code{\link{get_pgn_tags}()}
}
\concept{pgn file management tools}
