% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regan_antiderivative.R
\name{regan_antiderivative}
\alias{regan_antiderivative}
\title{Regan Antiderivative}
\usage{
regan_antiderivative(x)
}
\arguments{
\item{x}{An numeric vector of chess evaluations, in centipawns.}
}
\value{
A complex vector of the antiderivative of 1 / (1 + abs(x)).
}
\description{
Calculate the antiderivative of 1 / (1 + abs(x)).
}
\details{
This is a convenience function for calculating the definite
integral:

di = integral from v0 = eval(m0) to vi = eval(mi) of 1 / (1 + abs(x)) dx

used by Regan, et al., to scale the difference in the evaluation of an
engine's "best" move (m0) and a player's actual move (mi). The scaling
compensates for the human tendency to evaluate on a log scale.

The antiderivative of 1 / (1 + abs(x)) is:

1/2 * (sign(x) * (log(1 - x) + log(x + 1)) - log(1 - x) + log(x + 1))
}
\note{
The antiderivative is undefined at 1 or -1. At these points the
function will return the corresponding limit (+-log(4)/2).
}
\examples{
regan_antiderivative(c(-100, -10, -1, 0, 1, 10, 100))
}
\references{
\itemize{
\item Regan, K. W., & Haworth, G. M. C. (2011). Intrinsic Chess Ratings.
Twenty-Fifth AAAI Conference on Artificial Intelligence, 834–839.
https://www.aaai.org/ocs/index.php/AAAI/AAAI11/paper/view/3779
\item Regan, K. W., Macieja, B., & Haworth, G. M. C. (2012).
Understanding distributions of chess performances. Lecture Notes in
Computer Science, 7168, 230–243.
https://doi.org/10.1007/978-3-642-31866-5_20
}
}
