% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cram_positionlog.R
\name{remove_na_rows_cols}
\alias{remove_na_rows_cols}
\title{Remove rows and columns that are all NAs from a matrix}
\usage{
remove_na_rows_cols(mat)
}
\arguments{
\item{mat}{A matrix.}
}
\value{
A matrix with rows and columns that are all NAs removed.
}
\description{
Remove rows and columns that are all NAs from a matrix
}
