% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbitr_helpers.R
\name{scale_move_times}
\alias{scale_move_times}
\alias{inv_scale_move_times}
\title{Scale move times}
\usage{
scale_move_times(move_times)

inv_scale_move_times(scaled_times)
}
\arguments{
\item{move_times}{A numeric vector of move times.}

\item{scaled_times}{A numeric vector of scaled move times.}
}
\value{
For \code{scaled_move_times()}, a numeric vector of scaled move times. For
\code{inv_scaled_move_times()} a numeric vector of un-scaled move times.
}
\description{
Applies the scaling function log(0.005 * move_times + 3)^2 - log(3)^2 to
the \code{move_times} parameter.
}
\details{
The function \code{scale_move_times()} first applies a cap of 12e4, and
then applies the logarithmic scaling transformation
log(0.005 * \code{move_times} + 3)^2 - log(3)^2 to the \code{move_times} parameter.
The inverse function \code{inv_scale_move_times()} reverses the transformation.

The scaling function is borrowed from
\href{https://github.com/ornicar/lila/blob/442da0c86a9d54c3cff5645e14d67dfe269a9d0b/public/javascripts/chart/movetime.js}{lichess}
}
\note{
The inverse function cannot recover transformed values that were
greater than 12e4 due to the cap applied by the initial scaling.
}
