% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled_advantage_plot.R
\name{scaled_advantage_plot}
\alias{scaled_advantage_plot}
\title{Advantage plot}
\usage{
scaled_advantage_plot(scores, style = "graph")
}
\arguments{
\item{scores}{A numeric vector of positional evaluations (in centipawns).}

\item{style}{(Default = 'graph') A single-element character vector
indicating the plot style. Allowed values are 'graph' for a traditional
graph with axes, or 'infographic' to add a background gradient and remove
the axes (similar to \url{lichess.org}).}
}
\value{
A ggplot object of the plotted data.
}
\description{
Generate a plot of scaled evaluations (advantages).
}
\details{
Given a vector of positional evaluations (in centipawns), the
results are scaled using a form of
\href{https://lichess.org/blog/WFvLpiQAACMA8e9D/learn-from-your-mistakes}{logistic scaling function}
(2 / (1 + exp(-0.004 * scores)) - 1) borrowed from
\url{lichess.org}. This function downplays the impact of
less-than-perfect play when one side is far ahead. The idea being that the
player who is ahead may avoid moves that lead to material gain if those
moves lead to complicated tactics. In such positions players often go for
simple and solid moves that maintain a clear advantage. These types of
moves aren't really blunders. Similarly, a player in a lost position may
make moves designed to complicate the position and create chances for a
comeback, even if the move objectively loses material with perfect play by
the opponent. The scaled values are plotted vs. the number of half-moves.
}
\examples{
  scores <- c(15, 5, 29, -94, 67, 76, 154, -31, 1000, 1000)
  scaled_advantage_plot(scores)
}
\seealso{
\itemize{
\item \code{\link[=scaled_time_plot]{scaled_time_plot()}} to plot move time data.
\item \code{\link[=game_summary_plot]{game_summary_plot()}} to plot move time and advantage data with a
table of game stats.
}
}
