% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled_time_plot.R
\name{scaled_time_plot}
\alias{scaled_time_plot}
\title{Time plot}
\usage{
scaled_time_plot(white_move_times, black_move_times, style = "graph")
}
\arguments{
\item{white_move_times}{A numeric vector of move times for white, in seconds.}

\item{black_move_times}{A numeric vector of move times for black, in seconds.}

\item{style}{(Default = 'graph') A single-element character vector
indicating the plot style. Allowed values are 'graph' for a traditional
graph with axes, or 'infographic' to add a background gradient and remove
the axes (similar to \url{lichess.org}).}
}
\value{
A ggplot object of the plotted data.
}
\description{
Generate a plot of scaled move times.
}
\details{
Move times (in seconds) are plotted using a scaling function
borrowed from \url{lichess.org}. The
\href{https://github.com/ornicar/lila/blob/442da0c86a9d54c3cff5645e14d67dfe269a9d0b/public/javascripts/chart/movetime.js}{logarithmic scaling function}
keeps short move times from disappearing into the baseline. White move
times are shown as positive values, with a white fill, while black move
times are shown as negative values with a dark gray fill.
}
\examples{
  white_move_times <- c(4, 10, 5, 10)
  black_move_times <- c(3, 4, 7)
  scaled_time_plot(white_move_times, black_move_times)
}
\seealso{
\itemize{
\item \code{\link[=advantage_plot]{advantage_plot()}} to plot advantage data.
\item \code{\link[=lichess_plot]{lichess_plot()}} to plot move time and advantage data with a table
of game stats.
}
}
