% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_move.R
\name{search_move}
\alias{search_move}
\title{Search specific moves.}
\usage{
search_move(position, move, engine, depth)
}
\arguments{
\item{position}{A character vector of a chess position in long algebraic
notation (LAN).}

\item{move}{A character vector of moves to search.}

\item{engine}{An engine handler generated by the
\href{https://github.com/rosawojciech/bigchess}{bigchess} function
\code{uci_engine()}.}

\item{depth}{A single-element integer vector of the desired search depth, in
plies.}
}
\value{
A character vector of the engine output.
}
\description{
Search specific moves.
}
\examples{
library(bigchess)
# Modify the path as required for your engine location & operating system
engine <- bigchess::uci_engine('//stockfish_13_win_x64_bmi2.exe')
search_move('', 'h2h3', engine = engine, depth = 1)
uci_quit(engine)
}
