% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_moves.R
\name{search_moves}
\alias{search_moves}
\title{Search positions for the specified moves.}
\usage{
search_moves(positions, moves, engine, depth)
}
\arguments{
\item{positions}{A list of positions.}

\item{moves}{A list of moves.}

\item{engine}{An engine handler generated by the
\href{https://github.com/rosawojciech/bigchess}{bigchess} function
\code{uci_engine()}.}

\item{depth}{A single-element integer vector of the desired search depth, in
plies.}
}
\value{
A character vector of the engine output.
}
\description{
Given a list of positions and a list of moves, search each position using
only the moves supplied for that position. Useful when you want to know an
engine's evaluation for specific moves that might not be principal
variations.
}
\examples{
library(bigchess)
moves <- c("e2e4", "g7g5", "b1c3", "f7f5", "d1h5")
positions <- c( "", "e2e4", "e2e4 g7g5", "e2e4 g7g5 b1c3",
                "e2e4 g7g5 b1c3 f7f5")
# Modify the path as required for your engine location & operating system
engine <- uci_engine('//stockfish_13_win_x64_bmi2.exe')
search_moves(positions, moves, engine, depth = 1)
uci_quit(engine)
}
