% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_pgn.R
\name{search_pgn}
\alias{search_pgn}
\title{Search a pgn for games meeting the search criterion.
#'}
\usage{
search_pgn(pgn_path, pattern, n_max = -1)
}
\arguments{
\item{pgn_path}{A single-element character vector of the path to a pgn file.}

\item{pattern}{A single-element character vector containing the text to
search for.}

\item{n_max}{(Default = -1) A single-element integer vector of the maximum
number of matches to return. The default value of -1 returns all matches.}
}
\value{
A data frame of the games found.
}
\description{
Search a pgn for games meeting the search criterion.
#'
}
\details{
See the
\href{http://www.saremba.de/chessgml/standards/pgn/pgn-complete.htm#c8.1.1}{pgn specification}
for information on the pgn format.

The pgn will be searched for the specified text, and if found,
games containing the search text will be returned.
}
\examples{
pgn_path <- file.path(
  system.file(package = "rbitr"),
  'extdata',
  'fools_mate.pgn'
)
search_pgn(pgn_path, '1. e4 e5')
}
