% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_board.R
\name{update_board}
\alias{update_board}
\title{Update a chessboard to show the position after a move is made}
\usage{
update_board(move, board)
}
\arguments{
\item{move}{A single-element character vector consisting of a legal chess
move in UCI long algebraic notation.}

\item{board}{A list where the first element is a matrix representing the
chessboard, and the remaining 5 elements are the same as in a FEN string.}
}
\value{
A list containing the updated board and status. The list names should
be board, to_move, castling_rights, ep_target, halfmove_clock, and
fullmove_number.
}
\description{
This function updates a chessboard to show the position after a move is
made. It returns a list containing the updated board and the status of the
game. See \code{\link[=fen_to_board]{fen_to_board()}} for details of the board format.
}
\examples{
board <- fen_to_board()
move <- 'e2e4'
update_board(move, board)
}
