% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbitr_helpers.R
\name{winning_chances}
\alias{winning_chances}
\alias{inv_winning_chances}
\title{Calculate winning chances}
\usage{
winning_chances(scores)

inv_winning_chances(scaled_scores)
}
\arguments{
\item{scores}{A numeric vector of chess engine evaluations.}

\item{scaled_scores}{A numeric vector of scaled chess engine evaluations.}
}
\value{
A numeric vector of winning chances.
}
\description{
Applies the scaling function 2 / (1 + exp(-0.004 * \code{scores})) - 1 to the
\code{scores} parameter.
}
\details{
The function \code{winning_chances()} applies the exponential scaling
function 2 / (1 + exp(-0.004 * \code{scores})) - 1 to the \code{scores} parameter.
The inverse function \code{inv_winning_chances()} reverses the transformation.

The scaling function is borrowed from
\href{https://github.com/ornicar/lila/blob/442da0c86a9d54c3cff5645e14d67dfe269a9d0b/public/javascripts/chart/acpl.js}{lichess}
}
