% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_to_fen.R
\name{board_to_fen}
\alias{board_to_fen}
\title{Convert a chessboard matrix and status list to a FEN string}
\usage{
board_to_fen(board, status)
}
\arguments{
\item{board}{An 8x8 matrix representing a chess board. Each character
represents a piece (e.g., 'r' for black rook, 'R' for white rook, etc.), and
empty squares are represented by empty strings ('').}

\item{status}{A list containing the active color ('w' or 'b'), castling
availability, en passant target square, halfmove clock, and fullmove number.}
}
\value{
A FEN string representing the board.
}
\description{
The function \code{board_to_fen()} creates a Forsyth-Edwards Notation (FEN) string
from an 8x8 matrix of a chessboard and a list containing other information to
be stored in the FEN string. See
\href{http://www.saremba.de/chessgml/standards/pgn/pgn-complete.htm#c16.1}{section 16.1 of the PGN specification}
for detailed information about the FEN format. Also see the documentation for
\code{\link[=fen_to_board]{fen_to_board()}} for details about the board format.
}
\examples{
board <- fen_to_board()
status <- board[2:6]
board <- board[[1]]
board_to_fen(board, status)
}
\seealso{
\itemize{
\item \code{\link[=fen_to_board]{fen_to_board()}} To convert a FEN string to a board matrix.
\item \code{\link[=print_board]{print_board()}} Prints the board with rows in the correct order.
}
}
