% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_scores.R
\name{convert_scores}
\alias{convert_scores}
\title{Convert parsed score values to numeric scores.}
\usage{
convert_scores(scores, mate = 5000)
}
\arguments{
\item{scores}{A character vector of scores from a
\href{http://wbec-ridderkerk.nl/html/UCIProtocol.html}{UCI compatible} chess
engine.}

\item{mate}{(Default = 5000) A single-element integer or numeric vector of
the value to assign for 'mate in x'.}
}
\value{
An integer vector of scores (in centipawns).
}
\description{
Convert parsed score values to numeric scores.
}
\details{
The scores obtained from parsing the output of a
\href{http://wbec-ridderkerk.nl/html/UCIProtocol.html}{UCI compatible} chess
engine are a mix of centipawn scores, mate in x, and/or bounds on the score.
In addition, the scores for white will have opposite sign from the scores for
black. The function \code{convert_scores()} will convert bounds or mate in x into
numeric values. It will also change the sign of black's scores, so that
positive evaluations will mean white is ahead, and negative will mean black
is ahead.
}
\examples{
scores <- c("90", "-26", "26 upperbound", "mate -2" , "mate 1", "mate -1", "mate 0", NA)
convert_scores(scores)
}
