% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_gamelog.R
\name{cram_gamelog}
\alias{cram_gamelog}
\title{Deprecated alias for \code{\link[=tabulate_gamelog]{tabulate_gamelog()}}}
\usage{
cram_gamelog(
  gamelog,
  all_tags = FALSE,
  custom_tags = NULL,
  delete_blank_lines = TRUE
)
}
\arguments{
\item{gamelog}{A list that contains vectors of engine analysis.}

\item{all_tags}{(Default = \code{FALSE}) A Boolean. Setting \code{all_tags} = TRUE will
search for all of the tags listed in the UCI protocol.}

\item{custom_tags}{(Default = \code{NULL}) An optional character vector of custom
tag names.}

\item{delete_blank_lines}{(Default = \code{TRUE}) A Boolean. Setting this value to
FALSE will leave blank rows/columns intact.}
}
\value{
A data frame summarizing the data for the game.
}
\description{
The \code{tabulate_gamelog()} function takes a list of vectors of engine analysis
for a chess game and condenses it into a data frame for easier access.
}
\details{
The function \code{tabulate_gamelog()} parses output from a
\href{https://github.com/fsmosca/UCIChessEngineProtocol}{UCI-compatible} chess
engine.

\code{tabulate_gamelog()} is a wrapper for \code{tabulate_positionlog()} that
loops over each position in the game, and adds a column indicating the
position number. The \code{gamelog} parameter should contain output from the
\code{evaluate_game()} function. The remaining parameters are passed to
\code{tabulate_positionlog()}. See the documentation for
\code{\link[=tabulate_positionlog]{tabulate_positionlog()}} for details.
}
\examples{
gamelog <- list(c(
  "Stockfish 13 by the Stockfish developers (see AUTHORS file)",
  "readyok",
  "info string NNUE evaluation using nn-62ef826d1a6d.nnue enabled",
  "info depth 1 seldepth 1 multipv 1 score cp 146 nodes 30 nps 30000 tbhits 0 time 1 pv d2d4",
  "bestmove d2d4"
))
tabulate_gamelog(gamelog)
}
\seealso{
The 'tabulate' functions tabulate analysis logs into data frames.
\itemize{
\item \code{\link[=tabulate_positionlog]{tabulate_positionlog()}} for condensing analysis of one position.
\item \code{\link[=tabulate_gamelog]{tabulate_gamelog()}} for condensing analysis of one game.
\item \code{\link[=tabulate_pgnlog]{tabulate_pgnlog()}} for condensing analysis of an entire pgn file.
}

The 'parse' functions extract specific data from analysis logs.
\itemize{
\item \code{\link[=parse_gamelog]{parse_gamelog()}} for extracting data from one evaluated game.
\item \code{\link[=parse_pgnlog]{parse_pgnlog()}} for extracting data from games in a pgn.
}

The 'evaluate' functions produce analysis logs.
\itemize{
\item \code{\link[=evaluate_position]{evaluate_position()}} for analyzing chess positions.
\item \code{\link[=evaluate_game]{evaluate_game()}} for analyzing individual games.
\item \code{\link[=evaluate_pgn]{evaluate_pgn()}} for evaluating all the games in a PGN file.
}
}
