% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_pgn.R
\name{evaluate_pgn}
\alias{evaluate_pgn}
\title{Analyze each chess game in a PGN file}
\usage{
evaluate_pgn(
  pgn_path,
  engine_path,
  limiter,
  limit,
  n_cpus = 1L,
  n_pv = 1L,
  hash_size = NULL,
  mute = FALSE,
  save_logs = FALSE,
  save_path = NULL
)
}
\arguments{
\item{pgn_path}{A single-element character vector of the path to the PGN
file.}

\item{engine_path}{A single-element character vector of the path to a UCI
chess engine.}

\item{limiter}{A single-element character vector indicating the desired
mode of search termination. Allowed values are 'depth' (to search a fixed
number of plies), 'nodes' (to search a fixed number of nodes), and
'movetime' (to search a fixed number of milliseconds).}

\item{limit}{A single-element integer vector of the desired search depth
(# of plies), search nodes (# of nodes), or search time (# of mseconds).}

\item{n_cpus}{(Default = 1) A single-element integer vector of the number of
cpus to use.}

\item{n_pv}{(Default = 1) A single-element integer vector of the desired
number of principal variations.}

\item{hash_size}{(Default = NULL) A single-element integer vector of the
desired hash size, in MB.}

\item{mute}{(Boolean, default = FALSE) Suppress progress report?}

\item{save_logs}{(Boolean, default = FALSE) Save progress? Recommended for
long analyses in case it has to be stopped before finishing all games.}

\item{save_path}{A single-element character vector of the path to a directory
where the game logs should be saved.}
}
\value{
A list of gamelogs (see \code{evaluate_game()} for details).
}
\description{
The \code{evaluate_pgn()} function analyzes chess games in a PGN file using a
UCI-compatible chess engine. It applies the \code{evaluate_game()} function from
the \code{rbitr} package to each game in the PGN file and returns a list of game
logs.
}
\details{
Chess games are analyzed using a
\href{http://wbec-ridderkerk.nl/html/UCIProtocol.html}{UCI compatible} chess
engine.

The function \code{evaluate_pgn()} is a wrapper for rbitr's
\code{evaluate_game()} function that applies \code{evaluate_game()} to each game in
the specified PGN file. Note that this can take a very long time for PGN
files with a lot of games, and even longer when searching high numbers of
principal variations or running deep evaluations. For long runs, it is
recommended to set save_logs = TRUE. This will save the evaluation of each
game to a folder with the same name as the PGN file. If the analysis has to
be stopped and restarted, it will pick up where it left off if the
save_logs option was activated.

Since analysis can take a long time, progress is reported as each
game is completed, along with a rough estimate of the time remaining.
Progress reports may be turned off by setting mute = TRUE.

See \code{\link[=evaluate_game]{evaluate_game()}} for further details.
}
\note{
The server analysis on lichess.org uses a limit of 2250000 nodes. To
mimic this, set limiter = 'nodes', and limit = 2250000.
}
\examples{
# Modify engine_path as required for your engine location & operating system
engine_path <- '//stockfish.exe'
pgn_path <- file.path(
  system.file(package = 'rbitr'),
  'extdata',
  'fools_mate.pgn'
)
evaluate_pgn(pgn_path, engine_path, limiter = 'depth', limit = 1)
}
\seealso{
The 'tabulate' functions tabulate analysis logs into data frames.
\itemize{
\item \code{\link[=tabulate_positionlog]{tabulate_positionlog()}} for condensing analysis of one position.
\item \code{\link[=tabulate_gamelog]{tabulate_gamelog()}} for condensing analysis of one game.
\item \code{\link[=tabulate_pgnlog]{tabulate_pgnlog()}} for condensing analysis of an entire pgn file.
}

The 'parse' functions extract specific data from analysis logs.
\itemize{
\item \code{\link[=parse_gamelog]{parse_gamelog()}} for extracting data from one evaluated game.
\item \code{\link[=parse_pgnlog]{parse_pgnlog()}} for extracting data from games in a pgn.
}

The 'evaluate' functions produce analysis logs.
\itemize{
\item \code{\link[=evaluate_position]{evaluate_position()}} for analyzing chess positions.
\item \code{\link[=evaluate_game]{evaluate_game()}} for analyzing individual games.
\item \code{\link[=evaluate_pgn]{evaluate_pgn()}} for evaluating all the games in a PGN file.
}
}
