% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_position.R
\name{evaluate_position}
\alias{evaluate_position}
\title{Evaluate a chess position}
\usage{
evaluate_position(
  position,
  engine_path,
  limiter,
  limit,
  n_cpus,
  n_pv,
  hash_size = NULL
)
}
\arguments{
\item{position}{A single-element character vector containing a series of
legal chess moves in long algebraic notation (LAN).}

\item{engine_path}{A single-element character vector of the path to a UCI
compatible chess engine.}

\item{limiter}{A single-element character vector indicating the desired
mode of search termination. Allowed values are 'depth' (to search a fixed
number of plies), 'nodes' (to search a fixed number of nodes), and
'movetime' (to search a fixed number of milliseconds).}

\item{limit}{A single-element integer vector of the desired search depth
(# of plies), search nodes (# of nodes), or search time
(# of milliseconds).}

\item{n_cpus}{(Default = 1) A single-element integer vector of the number of
cpus to use.}

\item{n_pv}{(Default = 1) A single-element integer vector of the desired
number of principal variations.}

\item{hash_size}{(Default = NULL) A single-element integer vector of the
desired hash size, in MB.}
}
\value{
A list containing a character vector of the engine's output.
}
\description{
Chess positions are analyzed using a
\href{http://wbec-ridderkerk.nl/html/UCIProtocol.html}{UCI compatible} chess
engine.
}
\examples{
position <- 'e2e4'
# Modify engine_path as required for your engine location & operating system
engine_path <- '//stockfish.exe'
evaluate_position(position, engine_path, n_pv = 1, n_cpus = 1,
                  limiter = 'depth', limit = 1)
}
