% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_summary_plot.R
\name{game_summary_plot}
\alias{game_summary_plot}
\title{Game summery plot}
\usage{
game_summary_plot(
  pgn_path,
  game_number,
  engine_path = NULL,
  n_cpus = 1,
  use_pgn_evals = TRUE,
  nodes = 2250000,
  style = "graph"
)
}
\arguments{
\item{pgn_path}{A single-element character vector of the path to a PGN file.}

\item{game_number}{A single-element integer vector indicating which game in
the PGN file to plot.}

\item{engine_path}{A single-element character vector of the path to a UCI
compatible chess engine.}

\item{n_cpus}{(Default = 1) A single-element integer vector indicating how
many cpus to use for analysis.}

\item{use_pgn_evals}{(Default = TRUE) A single-element boolean indicating
whether to use evals from the PGN file, if present.}

\item{nodes}{(Default = 2250000) A single-element integer vector of the depth
to analyze to, in nodes.}

\item{style}{(Default = 'graph') A single-element character vector
indicating the plot style. Allowed values are 'graph' for a traditional
graph with axes, or 'infographic' to add a background gradient and remove
the axes (similar to \url{lichess.org}).}
}
\value{
A ggplot object of the plotted data.
}
\description{
Generate a plot of various game statistics.
}
\details{
The function \code{game_summary_plot()} will first look for existing
analysis in the PGN file, and if not found, it will look for a saved
analysis. If no analysis data is found, it will analyze the game. The
default search depth is 2,250,000 nodes.

An advantage plot and move time plot will be generated. (See
\code{advantage_plot()} and \code{scaled_time_plot()} for details.) In addition, a
table will be displayed showing the number of inaccuracies, mistakes, and
blunders, and the average centipawn loss (ACPL). (See \code{get_imb()} and
\code{get_acpl()}.) If clock data is not available in the PGN file, the
move-time plot will display a message that no data is available.
}
\examples{
pgn_path <- file.path(
  system.file(package = 'rbitr'),
  'extdata',
  'short_game.pgn'
)
# Modify engine_path as required for your engine location & operating system
engine_path <- '//stockfish_13_win_x64_bmi2.exe'
game_summary_plot(pgn_path, game_number = 1, engine_path)
}
\seealso{
\itemize{
\item \code{\link[=scaled_advantage_plot]{scaled_advantage_plot()}} to plot advantage data.
\item \code{\link[=scaled_time_plot]{scaled_time_plot()}} to plot move time data.
\item \code{\link[=get_acpl]{get_acpl()}} to calculate average centipawn loss.
\item \code{\link[=get_imb]{get_imb()}} to calculate inaccuracies, mistakes, and blunders.
}
}
