% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acpl.R
\name{get_acpl}
\alias{get_acpl}
\title{Calculate the average centipawn loss.}
\usage{
get_acpl(scores, color, cap = NULL, cap_action = "none", first_ply = 1)
}
\arguments{
\item{scores}{A numeric or integer vector of evaluations, in centipawns.}

\item{color}{A single-element character vector indicating which side to
evaluate. Allowed values are 'black' or 'white'.}

\item{cap}{(Default = NULL) A single-element numeric or integer vector
indicating the threshold for capping a move's loss (or excluding the move
from the average).}

\item{cap_action}{(Default = NULL) A single-element character vector
indicating whether to exclude scores outside the range [-cap, cap], or to
replace them with the cap. Allowed values are 'exclude', 'replace', or
'none'.}

\item{first_ply}{(Default = 1) A single-element integer vector indicating the
first ply to be included in the calculation. May not be larger than  the
number of elements in \code{score}}
}
\value{
A single-element integer vector of the average centipawn loss.
}
\description{
Calculate the average centipawn loss.
}
\details{
Calculates the average centipawn loss (ACPL) for the side indicated
by the \code{color} parameter. The \code{score} parameter is assumed to begin with a
score for the initial position, prior to the first move. Note that when
pgn files store evaluation scores, the score for the initial position is
omitted.

The method for calculating average centipawn loss varies.
In Guid & Bratko, 'MeanLoss' is calculated only after excluding
the first 12 moves, and also excluding moves where both the best move and
the played move result in evaluations outside the range [-200, 200]. To
replicate MeanLoss, set \code{cap} to 200, \code{cap_action} to 'exclude' and,
\code{first_ply} to 25.

The website lichess.org uses a different method, in which
evaluations outside the range [-1000, 1000] are replaced with +-1000, and
the initial position is always given a score of 15. To replicate the
lichess ACPL, set \code{cap} to 1000, \code{cap_action} to 'replace', and make sure
that the first element of \code{scores} is equal to 15.
}
\note{
To calculate ACPL for games that start at a position other than the
initial position, the \code{score} parameter may begin at a later position,
provided that it begins with a position where white is to move.
}
\examples{
scores <- c(5, 29, -94, 67, 76, 154, -31, 1000, 1000)
get_acpl(scores, 'white', first_ply = 2)
}
\references{
M. Guid, I. Bratko. Computer analysis of world chess champions.
ICGA Journal, 29(2):65–73, 2006.
}
