% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_movetext.R
\name{get_clocks}
\alias{get_clocks}
\alias{get_evals}
\title{Get command tags from pgn movetext}
\usage{
get_clocks(movetext)

get_evals(movetext, first_eval = NULL, mate0 = FALSE, mate_value = 50000)
}
\arguments{
\item{movetext}{A character vector of pgn movetext, where each vector entry
is for a separate game.}

\item{first_eval}{(Default = NULL) A single-element integer vector indicating
what value should be assigned to the initial position, before white's first
move. The default (NULL) omits the first evaluation.}

\item{mate0}{(Default = FALSE) A single-element boolean vector indicating
whether to include a value for mated positions.}

\item{mate_value}{(Default = 50000) A single-element integer or numeric
vector of the centipawn value to assign for mate.}
}
\value{
A list containing numeric vectors of the specified tag value. Each
list entry is for a separate game. For \code{get_clocks()}, the value will be
converted to seconds.
}
\description{
Get the specified command tag (if present) from the movetext taken from a
pgn file.
}
\details{
See the
\href{http://www.enpassant.dk/chess/palview/enhancedpgn.htm}{Portable Game Notation Specification and Implementation Guide (Supplement)}
for details on command tags in pgn files.

The \code{get_clocks()} and \code{get_evals()} functions are intended for use
with data read from a pgn using the \code{get_pgn()} function. A tibble obtained
using \code{get_pgn()} will have a $Movetext column from which the tag data can
be extracted if it is present.
}
\note{
For games stored in pgn format, the evaluation of the first position
before any moves are made is not commonly included. As a result,
\code{get_evals()} produces a vector that is one evaluation shorter than the
one produced by \code{parse_gamelog()} for the same game. To include a value for
the initial position, set the parameter first_eval to the desired value.
To use the same default as \href{http://lichess.org}{lichess.org}, set
first_eval to 15.

For games ending in mate, a uci engine gives the final position an
evaluation of 'mate 0', however, in a pgn, any '#x' notation is omitted for
the final position when it is mate. When the final position is mate,
\href{http://lichess.org}{lichess.org} advantage plots show an advantage of 1
for the final position, but the value is not used in calculating acpl. To
include an evaluation for mated positions, set the mate0 parameter to TRUE.
}
\examples{
movetext <- c(
  '1. d4 {[\%clk 0:15:00]} d5 {[\%clk 0:15:00]} 2. Nc3 {[\%clk 0:14:48]}',
  '1. e4 {[\%clk 0:15:00]} e5 {[\%clk 0:15:00]} 2. Nf3 {[\%clk 0:14:38]}'
)
get_clocks(movetext)

movetext <- c(
  '1. e4 { [\%eval 0.05] } 1... Nf6 { [\%eval 0.29] } 2. Bc4 { [\%eval -0.94] }'
)
get_evals(movetext)
}
