% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgn_data_manipulators.R
\name{get_clocks}
\alias{get_clocks}
\title{Get clock data from pgn movetext}
\usage{
get_clocks(source_movetext)
}
\arguments{
\item{source_movetext}{A character vector of pgn movetext.}
}
\value{
A list containing numeric vectors of time remaining after each move,
in seconds.
}
\description{
Extract clock data (if present) from the movetext of a pgn file.
}
\details{
See http://www.enpassant.dk/chess/palview/enhancedpgn.htm for the
specification of the clk command in pgn files.

For pgns read with \code{bigchess::read.pgn} or \code{rbitr::read_pgn}, the
Movetext column is stripped of embedded commands and will not contain clock
data. The original, unmodified movetext is found in the SourceMovetext
column.
}
\examples{
source_movetext <- c(
  '1. d4 {[\%clk 0:15:00]} d5 {[\%clk 0:15:00]} 2. Nc3 {[\%clk 0:14:48]}',
  '1. e4 {[\%clk 0:15:00]} e5 {[\%clk 0:15:00]} 2. Nf3 {[\%clk 0:14:38]}'
)
get_clocks(source_movetext)
}
\seealso{
Other pgn data manipulators: 
\code{\link{calculate_move_times}()},
\code{\link{get_increments}()}
}
\concept{pgn data manipulators}
