% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_movetext.R
\name{get_clocks}
\alias{get_clocks}
\alias{get_evals}
\title{Get command tags from pgn movetext}
\usage{
get_clocks(movetext)

get_evals(movetext)
}
\arguments{
\item{movetext}{A character vector of pgn movetext, where each vector entry
is for a separate game.}
}
\value{
A list containing numeric vectors of the specified tag value. Each
list entry is for a separate game. For \code{get_clocks()}, the value will be
converted to seconds.
}
\description{
Get the specified command tag (if present) from the movetext taken from a
pgn file.
}
\details{
See the
\href{http://www.enpassant.dk/chess/palview/enhancedpgn.htm}{Portable Game Notation Specification and Implementation Guide (Supplement)}
for details on command tags in pgn files.

The \code{get_clocks()} and \code{get_evals()} functions are intended for use
with data read from a pgn using the \code{get_pgn()} function. A tibble obtained
using \code{get_pgn()} will have a $Movetext column from which the tag data can
be extracted if it is present.
}
\examples{
movetext <- c(
  '1. d4 {[\%clk 0:15:00]} d5 {[\%clk 0:15:00]} 2. Nc3 {[\%clk 0:14:48]}',
  '1. e4 {[\%clk 0:15:00]} e5 {[\%clk 0:15:00]} 2. Nf3 {[\%clk 0:14:38]}'
)
get_clocks(movetext)

movetext <- c(
  '1. e4 { [\%eval 0.05] } 1... Nf6 { [\%eval 0.29] } 2. Bc4 { [\%eval -0.94] }'
)
get_evals(movetext)
}
