% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgn_data_manipulators.R
\name{get_increments}
\alias{get_increments}
\title{Get increments from pgn TimeControl}
\usage{
get_increments(time_control)
}
\arguments{
\item{time_control}{A character vector of pgn TimeControl tags.}
}
\value{
An integer vector of the increments, in seconds. Time controls with
no increment return 0.
}
\description{
Extract the increment from a pgn time control tag. For allowed time control
formats, see:
http://www.saremba.de/chessgml/standards/pgn/pgn-complete.htm#c9.6
}
\examples{
time_control <- c('?', '-', '40/9000', '300', '4500+60', '*180')
get_increments(time_control)
}
\seealso{
Other pgn data manipulators: 
\code{\link{calculate_move_times}()},
\code{\link{get_clocks}()}
}
\concept{pgn data manipulators}
