% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_move_times.R
\name{get_move_times}
\alias{get_move_times}
\title{Convert clock data to move times}
\usage{
get_move_times(clock, increment, color, negative_time = "NA")
}
\arguments{
\item{clock}{A numeric vector of chess clock times at the end of each half
move (in seconds).}

\item{increment}{A single-element numeric vector of the increment time (in
seconds).}

\item{color}{A character vector indicating which side to calculate (either
black or white).}

\item{negative_time}{(Default = 'NA') A character vector indicating how to
treat negative move times. If negative_time = 'NA', negative times will
be replaced with NAs. A value of 'no action' will leave negative move times
in place.}
}
\value{
A numeric vector of the move times in seconds. The vector starts at
the second move (the first move would count as zero seconds).
}
\description{
Clock data (in seconds) is converted into move times (in seconds).
}
\details{
Clock times and increments are typically obtained using rbitr's
\code{get_clocks()} and \code{get_increments()} functions. The move times start from
the second move (the first move is taken to be zero seconds).
}
\note{
The author has encountered PGN files where a move resulted in the
remaining time increasing by more than the increment value, giving a
negative move times. This might occur due to a software bug causing the
time to be incorrectly recorded, or perhaps by the action of an arbiter
(for example, when adding time to an opponents clock as a penalty). If
\code{get_move_times()} finds a negative move time, it will take action as
directed by the \code{negative_time} parameter. A value of 'NA' will treat
negative move times as missing data, replacing them with NAs. A value of
'no_action' will leave negative move times in place, treating them as
unrecorded additions of time to the clock.
}
\examples{
clock <- c(900, 900, 888, 878, 878, 858)
get_move_times(clock, 8, 'white')
get_move_times(clock, 8, 'black')
}
\seealso{
\itemize{
\item \code{\link[=get_pgn]{get_pgn()}} to load the time control and other data from a PGN
file.
\item \code{\link[=get_clocks]{get_clocks()}} to extract the clock data from a PGN file.
\item \code{\link[=get_increments]{get_increments()}} to extract increment data from a PGN file.
\item \code{\link[=time_plot]{time_plot()}} and \code{\link[=game_summary_plot]{game_summary_plot()}} to plot move
times.
}
}
