% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_moves.R
\name{get_moves}
\alias{get_moves}
\title{Gets LAN format moves from a pgn movetext vector}
\usage{
get_moves(movetext)
}
\arguments{
\item{movetext}{A character vector of one or more movetext fields from a pgn.}
}
\value{
A list of character vectors, where each vector contains the LAN
format moves of the corresponding movetext.
}
\description{
Given a character vector of movetext (typically obtained from a pgn using
rbitr's \code{get_pgn()} function), return a list of character vectors of LAN
format moves.
}
\examples{
movetext <- c('1. e4 e5', '1. h3 h6')
get_moves(movetext)
}
