% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pgn.R
\name{get_pgn}
\alias{get_pgn}
\title{Get the contents of a pgn file.}
\usage{
get_pgn(pgn_path)
}
\arguments{
\item{pgn_path}{A single-element character vector of the path to the pgn.}
}
\value{
A tibble where each row is a game, and each pgn tag is a column.
The pgn's movetext field is also a column.
}
\description{
Get the contents of a pgn file.
}
\details{
A no-frills function to read a pgn file into a tibble. Every tag
in the  pgn will be read.

See the
\href{http://www.saremba.de/chessgml/standards/pgn/pgn-complete.htm#c8.1.1}{pgn specification}
for information on the pgn format.
}
\examples{
pgn_path <- file.path(
  system.file(package = 'rbitr'),
  'extdata',
  'fools_mate.pgn'
)
get_pgn(pgn_path)
}
