% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_positions.R
\name{get_positions}
\alias{get_positions}
\title{Generate positions from movetext}
\usage{
get_positions(movetext)
}
\arguments{
\item{movetext}{A string containing chess moves in SAN.}
}
\value{
A character vector of chess positions.
}
\description{
This function takes a string containing chess moves in standard algebraic
notation (SAN) and returns a character vector of positions, also in SAN. The
initial position is represented by an empty string "". Subsequent positions
are represented by the series of chess moves that generated the position.
This is mainly useful for generating positions that can be fed to a chess
engine for analysis.
}
\details{
The function \code{get_positions} expects the input to be valid SAN that conforms
to the
\href{http://www.saremba.de/chessgml/standards/pgn/pgn-complete.htm}{PGN specification}.
It does not perform any checks to ensure that the input is valid and consists
of legal chess moves. Annotations, game termination markers, comments, etc.,
will be removed from the output.
}
\examples{
movetext <- "1. e4 e5 2. Nf3 Nc6 3. Bb5 a6"
positions <- get_positions(movetext)
print(positions)
}
