% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_columns.R
\name{harmonize_columns}
\alias{harmonize_columns}
\title{Harmonize columns in a list of data frames}
\usage{
harmonize_columns(df_list)
}
\arguments{
\item{df_list}{A list of data frames}
}
\value{
A list of data frames with harmonized columns
}
\description{
This function takes a list of data frames and adds missing columns to each
data frame so that all data frames have the same columns in the same order.
}
\details{
Any added columns will be populated with NAs. The function can be
used to prepare data frames for an rbind when the columns are not identical.
}
\examples{
# Create some example data frames
df1 <- data.frame(a = 1:3, b = letters[1:3])
df2 <- data.frame(b = letters[4:6], c = 4:6)
df3 <- data.frame(a = 7:9, c = letters[7:9])

# Harmonize columns
harmonize_columns(list(df1, df2, df3))
}
