% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_check.R
\name{is_check}
\alias{is_check}
\title{Look for checks from enemy pieces}
\usage{
is_check(board, color)
}
\arguments{
\item{board}{An 8x8 matrix representing a chess position. See
\code{\link[=fen_to_board]{fen_to_board()}} for details of the board representation.}

\item{color}{A string ("white" or "black"), indicating which king to examine.}
}
\value{
TRUE if the king is in check from an enemy piece, FALSE otherwise.
}
\description{
This function returns TRUE if there are checks from enemy pieces, and FALSE
otherwise.
}
\examples{
board <- fen_to_board('8/8/8/8/4k3/8/8/8 b - - 0 1')
print(board$board)
is_check(board$board, 'black')
board$board[3, 4] <- 'P'
print(board$board)
is_check(board$board, 'black')
}
