% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lichess_advantage_plot.R
\name{lichess_advantage_plot}
\alias{lichess_advantage_plot}
\title{Lichess advantage plot}
\usage{
lichess_advantage_plot(scores)
}
\arguments{
\item{scores}{A numeric vector of positional evaluations (in centipawns.)}
}
\value{
A ggplot object of the plotted data.
}
\description{
Generate an advantage plot in the style of \url{lichess.org}.
}
\details{
Given a vector of positional evaluations (in centipawns), the
results are first scaled using an
\href{https://lichess.org/blog/WFvLpiQAACMA8e9D/learn-from-your-mistakes}{exponential scaling function}
(2 / (1 + exp(-0.004 * scores)) - 1), and then plotted using the same color
scheme as \url{lichess.org}.
}
\examples{
  scores <- c(15, 5, 29, -94, 67, 76, 154, -31, 1000, 1000)
  lichess_advantage_plot(scores)
}
\seealso{
\itemize{
\item \code{\link[=scaled_time_plot]{scaled_time_plot()}} to plot move time data.
\item \code{\link[=lichess_plot]{lichess_plot()}} to plot move time and advantage data with a table
of game stats.
}
}
