% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lichess_time_plot.R
\name{lichess_time_plot}
\alias{lichess_time_plot}
\title{Lichess time plot}
\usage{
lichess_time_plot(white_move_times, black_move_times)
}
\arguments{
\item{white_move_times}{A numeric vector of move times for white, in seconds.}

\item{black_move_times}{A numeric vector of move times for black, in seconds.}
}
\value{
A ggplot object of the plotted data.
}
\description{
Generate a plot of move times in the style of
\href{http://lichess.org}{lichess.org}.
}
\details{
Move times (in seconds) are plotted using the same scaling and
color scheme used by \href{http://lichess.org}{lichess.org}.
}
\examples{
  white_move_times <- c(4, 10, 5, 10)
  black_move_times <- c(3, 4, 7)
  lichess_time_plot(white_move_times, black_move_times)
}
