% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lichess_time_plot.R
\name{lichess_time_plot}
\alias{lichess_time_plot}
\title{Lichess time plot}
\usage{
lichess_time_plot(white_move_times, black_move_times)
}
\arguments{
\item{white_move_times}{A vector of move times for white, in seconds.}

\item{black_move_times}{A vector of move times for black, in seconds.}
}
\value{
A ggplot object of the plotted data.
}
\description{
#' Generate a move times plot similar to the ones in the "Computer analysis"
tab on lichess.org.
}
\details{
Given vectors of move times (in seconds), the results are scaled
using the same logarithmic function used by lichess.org, and the results
are plotted using the same color scheme.
}
\examples{
  white_move_times <- c(4, 10, 5, 10)
  black_move_times <- c(3, 4, 7)
  lichess_time_plot(white_move_times, black_move_times)
}
