% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gamelog.R
\name{parse_gamelog}
\alias{parse_gamelog}
\title{Extract data from rbitr's \code{evaluate_game()} output}
\usage{
parse_gamelog(gamelog, target, depth = NULL)
}
\arguments{
\item{gamelog}{A list of engine output from rbitr's \code{evaluate_game()}
function.}

\item{target}{A single-element character vector of the output to return.
Allowed values are 'score' for the evaluation in centipawns, 'pv' for the
principal variation, or 'bestmove' for the best move.}

\item{depth}{(Optional, default = NULL) A single-element integer vector
indicating which search depth to return. The value must not be less than 1
or greater than the maximum depth reported by the engine. A value of NULL
returns data for the maximum depth.}
}
\value{
A list of character vectors of the extracted data, where each list
entry corresponds to a position.
}
\description{
The \code{evaluate_game()} function produces a list of engine output for each
position of a chess game. This output tends to be dense, and much of it
will be of no interest to someone analyzing a game. The \code{parse_gamelog()}
function can extract the parts of interest into a more manageable format.
}
\details{
Use the \code{target} parameter to specify what \code{parse_gamelog()} should return:
positional evaluations in centipawns (score), principal variations (pv), or
best moves (bestmove).

The data for the deepest available search depth is returned by
default. To extract data for shallower depths, set the \code{depth}
parameter to the desired depth. Note that if \code{evaluate_game()} was run
using fixed nodes or fixed time, the deepest search depth will vary across
different positions. If \code{depth} is set to a value higher than what is
available for some moves, \code{evaluate_game()} will return an error.

See
\href{http://wbec-ridderkerk.nl/html/UCIProtocol.html}{the UCI protocol} for
details of UCI engine output.
}
\examples{
gamelog <- list(c(
  "Stockfish 13 by the Stockfish developers (see AUTHORS file)",
  "readyok",
  "info string NNUE evaluation using nn-62ef826d1a6d.nnue enabled",
  "info depth 1 seldepth 1 multipv 1 score cp 146 nodes 30 nps 30000 tbhits 0 time 1 pv d2d4",
  "bestmove d2d4"
))
parse_gamelog(gamelog, 'score')
parse_gamelog(gamelog, 'pv')
parse_gamelog(gamelog, 'bestmove')
}
\seealso{
\itemize{
\item \code{\link[=evaluate_game]{evaluate_game()}} for analyzing individual games.
\item \code{\link[=evaluate_pgn]{evaluate_pgn()}} for evaluating all the games in a PGN file.
}
}
