% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pgnlog.R
\name{parse_pgnlog}
\alias{parse_pgnlog}
\title{Extract data from rbitr's \code{evaluate_pgn()} output}
\usage{
parse_pgnlog(pgnlog, target, depth = NULL)
}
\arguments{
\item{pgnlog}{A list of gamelogs from rbitr's \code{evaluate_pgn()} function.}

\item{target}{A single-element character vector of the output to return.
Allowed values are 'score' for the evaluation in centipawns, 'pv' for the
principal variation, or 'bestmove' for the best move.}

\item{depth}{(Optional, default = NULL) A single-element integer vector
indicating which search depth to return. The value must not be less than 1
or greater than the maximum depth reported by the engine. A value of NULL
returns data for the maximum depth.}
}
\value{
A list of vectors of the extracted data.
}
\description{
The \code{parse_pgnlog()} function is a wrapper for the \code{parse_gamelog()} function
that iterates over all the games in a PGN file.
}
\details{
The output of rbitr's \code{evaluate_pgn()} function is a list of logs
from the \code{evaluate_game()} function, where each entry in the list is for a
single game. The \code{parse_pgnlog()} function can extract  evaluations in
centipawns (score), principal variations (pv), or best moves (bestmove),
depending on the setting for the \code{target} parameter.

See \code{\link[=parse_gamelog]{parse_gamelog()}} for more details.
}
\examples{
pgnlog <- list(list(c(
"Stockfish 13 by the Stockfish developers (see AUTHORS file)",
"readyok",
"info string NNUE evaluation using nn-62ef826d1a6d.nnue enabled",
"info depth 1 seldepth 1 multipv 1 score cp 146 nodes 30 nps 30000 tbhits 0 time 1 pv d2d4",
"bestmove d2d4"
)))
parse_pgnlog(pgnlog, 'score')
parse_pgnlog(pgnlog, 'pv')
parse_pgnlog(pgnlog, 'bestmove')
}
\seealso{
\itemize{
\item \code{\link[=parse_gamelog]{parse_gamelog()}} for extracting data from one evaluated game.
\item \code{\link[=evaluate_game]{evaluate_game()}} for analyzing individual games.
\item \code{\link[=evaluate_pgn]{evaluate_pgn()}} for evaluating all the games in a PGN file.
}
}
